% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_functions.R
\name{addLogEntry}
\alias{addLogEntry}
\title{Add entry to analysis log}
\usage{
addLogEntry(entry.name, entry, df.log, entry.variable.name = "")
}
\arguments{
\item{entry.name}{name of entry. A character.}

\item{entry}{entry.}

\item{df.log}{existing analysis log to add entry to. A data.frame.}

\item{entry.variable.name}{name of variable storing entry (optional). A character.}
}
\value{
datalog data.frame
}
\description{
Add entry to analysis log
}
\examples{

# intiate data log
df.log <- initiateLog("log name")

# add new entry to log
df.log <- addLogEntry("Query File (.Rdata)", (input.file), df.log, "input.file")

}
\seealso{
\code{\link{initiateLog}}
}
\author{
Nicholas Mikolajewicz
}
