% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{AddSModuleScore}
\alias{AddSModuleScore}
\title{Calculate standardized module scores for feature expression programs in single cells.}
\usage{
AddSModuleScore(
  object,
  features,
  pool = NULL,
  nbin = NULL,
  ctrl = 100,
  nworkers = 1,
  k = FALSE,
  assay = NULL,
  name = "Cluster",
  seed = 1,
  search = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat Object}

\item{features}{A list of vectors of features for expression programs; each entry should be a vector of feature names}

\item{pool}{List of features to check expression levels against, defaults to `rownames(x = object)`}

\item{nbin}{Number of bins of aggregate expression levels for all analyzed features}

\item{ctrl}{Number of control features selected from the same bin per analyzed feature}

\item{nworkers}{Number of workers for parallel implementation. Default is 1.}

\item{k}{Use feature clusters returned from DoKMeans}

\item{assay}{Name of assay to use}

\item{name}{Name for the expression programs; will append a number to the end for each entry in features (eg. if `features` has three programs, the results will be stored as name1, name2, name3, respectively)}

\item{seed}{Set a random seed. If NULL, seed is not set.}

\item{search}{Search for symbol synonyms for features in features that don't match features in object? Searches the HGNC's gene names database; see UpdateSymbolList for more details}

\item{...}{Extra parameters passed to UpdateSymbolList}
}
\value{
Returns seurat object with standardized module scores added to object meta data; each module is stored as name# for each module program present in `features`
}
\description{
Calculates the standardized expression level of each program on single-cell level, by subtracting by aggregated expression of control feature sets (like in `AddModuleScore`) and then scaling the difference by the pooled variance of the gene set of interest and control features. Like `AddModuleScore`, all analyzed features are binned based on averaged expression and control features are randomly selected from each bin.
}
\seealso{
\code{\link{AddModuleScore}} for original module scoring function implemented in `Seurat`.
}
\author{
Nicholas Mikolajewicz
}
