% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{annotationCloud}
\alias{annotationCloud}
\title{Word cloud visualization of cell-type annotations from the Miko scoring pipeline}
\usage{
annotationCloud(
  object,
  object.group = "seurat_clusters",
  score,
  score.group,
  score.cell.type,
  score.p,
  score.fdr = NULL,
  score.coherence.fraction = NULL,
  score.frequent.flier = NULL,
  fdr.correction = T,
  p.threshold = 0.05,
  coherence.threshold = 0.8,
  show.n.terms = 15,
  verbose = T
)
}
\arguments{
\item{object}{Seurat Object.}

\item{object.group}{name of object meta data field specifying cluster membership. Default is "seurat_clusters".}

\item{score}{vector of Miko scores}

\item{score.group}{vector of group memberships}

\item{score.cell.type}{vector of cell-type names/labels.}

\item{score.p}{vector of p values.}

\item{score.fdr}{vector of fdr values. Optional.}

\item{score.coherence.fraction}{vector of coherence fractions. See coherentFraction(...) for details.}

\item{score.frequent.flier}{vector of logicals specifying whether score belongs to frequent flier.}

\item{fdr.correction}{Specify whether p-value should be corrected using Benjamini & Hochberg method. Default is T.}

\item{p.threshold}{p value threshold. Default is 0.05.}

\item{coherence.threshold}{Numerical [0,1] specifying minimal coherence required to qualify for visualization. Default is 0.8.}

\item{show.n.terms}{Maximal number of cell-type terms shown in word cloud. Default is 15.}

\item{verbose}{Logical, specify whether process is printed. Default is T.}
}
\value{
list of ggplot handles
}
\description{
Given outputs from the Miko scoring pipeline, the top cell-type annotations fare visualized using word clouds.
}
\examples{

df.score_summary <- data.frame(cluster = df.merge$cluster,
                              cell.type = df.merge$gs,
                              miko_score = signif(df.merge$miko_score, 3) ,
                              p =  signif(df.merge$p),
                              fdr =  signif(df.merge$fdr),
                              coherence_fraction =  signif(df.merge$coherence_fraction))


plt.cloud <- annotationCloud(object = so.query_scored,
                             object.group = "seurat_clusters",
                             score = df.score_summary$miko_score,
                             score.group = df.score_summary$cluster,
                             score.cell.type = df.score_summary$cell.type,
                             score.p = df.score_summary$p,
                             score.fdr = df.score_summary$fdr,
                             score.coherence.fraction = df.score_summary$coherence_fraction,
                             score.frequenct.flier = NULL,
                             fdr.correction = T,
                             p.threshold = 0.05,
                             coherence.threshold = 0.9,
                             show.n.terms = 15,
                             verbose = T)

}
\seealso{
\code{\link{mikoScore}} for miko scoring, \code{\link{coherentFraction}} for coherence scoring
}
\author{
Nicholas Mikolajewicz
}
