% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{autoPointSize}
\alias{autoPointSize}
\title{Automatically determine optimal point size for geom_point()}
\usage{
autoPointSize(n.points, scale.factor = 1583)
}
\arguments{
\item{n.points}{Number of data points}

\item{scale.factor}{scaling constant used to determine optimal point size.}
}
\value{
optimal point size
}
\description{
Automatically determine optimal point size for geom_point()
}
\examples{

 gg.plot <-  df.umap \%>\% dplyr::arrange(get(module.names[i])) \%>\%
      ggplot(aes(x = x, y = y, color = get(module.names[i]))) +
      geom_point(size = autoPointSize(nrow(df.umap)))

}
