% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{BaCo}
\alias{BaCo}
\title{Bayesian Correlation algorithm}
\usage{
BaCo(X)
}
\arguments{
\item{X}{Expression matrix}
}
\value{
Correlation matrix
}
\description{
Bayesian correlation scheme that assigns low similarity to genes that have low confidence expression estimates. Shown to be more reproducible than Pearson correlations. Source: https://www.biorxiv.org/content/10.1101/714824v1
}
\examples{

#create a matrix (or load your own)
X <- matrix(1:1000, ncol=20)

#compute the Bayesian correlation matrix
B <- BaCo(X)

}
