% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{baderPathways}
\alias{baderPathways}
\title{Pathway annotations from Bader Lab}
\format{
A list of pathways
}
\usage{
{
data(baderPathways)


# To update list:

# specify location of input data
# directory <- "D:/Users/Nick/Dropbox/PDF Projects - JM/Data/CRISPR screen/GIN Data/"
# file.pathways <- list(
#   Hs.entrez = "Human_GOBP_AllPathways_no_GO_iea_April_01_2020_entrezgene.gmt",
#   Hs.symbol = "Human_GOBP_AllPathways_no_GO_iea_April_01_2020_symbol.gmt",
#   Mm.entrez = "Mouse_GOBP_AllPathways_no_GO_iea_April_01_2020_entrezgene.gmt",
#   Mm.symbol = "Mouse_GOBP_AllPathways_no_GO_iea_April_01_2020_symbol.gmt"
# )
#
# pathway.list <- list()
# for (i in 1:length(file.pathways)){
#
#   df.pathway.2 <- qusage::read.gmt(paste0(directory, file.pathways[[names(file.pathways)[i]]]))
#   pathways <- df.pathway.2
#   names(pathways) <- gsub("%.*", "", names(pathways))
#
#   pathway.list[[names(file.pathways)[i]]] <- pathways
#
# }
#
# baderPathways <- pathway.list
#
# output.dir <- "D:/Users/Nick/Dropbox/PDF Projects - JM/R Packages/scMiko/data/"
#
# save(baderPathways, file=getLoadPath("baderPathways.rda", output.dir))
}
}
\description{
Updated list of gene sets used for enrichment analysis. Consolidated by Bader Lab, includes KEGG, Msigdb, NCI, IOB, NetPath, HumanCyc, Reactome, GO, Panther. http://baderlab.org/GeneSets
}
\examples{
data(baderPathways)

}
\keyword{annotation}
\keyword{genesets,}
\keyword{pathway}
