% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{balanceMatrixSize}
\alias{balanceMatrixSize}
\title{Balance matrix dimensions}
\usage{
balanceMatrixSize(mat.1, mat.2, method = "match.max")
}
\arguments{
\item{mat.1}{First input matrix.}

\item{mat.2}{Second input matrix.}

\item{method}{Character specifying which matching method to use. Must be one of "match.max" (default) or "match.min". For either option, matrix with min/max number of rows is resamples to match the other matrix.}

\item{s.mat}{similarity matrix}
}
\value{
list of resampled matrices.
}
\description{
Resamples two matrices to match number of rows in each. Number of rows is matched to minimum or maximum, as specified.
}
\examples{

# ensure sample sizes are balanced across groups
output.mat <- balanceMatrixSize(de.orig.1, de.orig.2, method = "match.max")
de.1 <- output.mat[["de.1"]] # resampled mat.1
de.2 <- output.mat[["de.2"]] # resampled mat.2
datExpr.noz <- output.mat[["de.all"]] # concatenated matrices

}
\author{
Nicholas Mikolajewicz
}
