% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{balanceSamples}
\alias{balanceSamples}
\title{Subsample cells in seurat object to be balanced (sample-size-wise) across conditions.}
\usage{
balanceSamples(object, group, balance.size = NA)
}
\arguments{
\item{object}{Seurat Object}

\item{group}{Character specifying group (i.e., metadata field) to sub-sample within.}

\item{balance.size}{Numeric. Target sub sample size for each condition. If target exceeds number of available cells, only available cells are included. If unspecified, smallest-sized group is set as target (Default).}
}
\value{
Seurat object
}
\description{
Subsample cells in seurat object to be balanced (sample-size-wise) across conditions.
}
\examples{

so.query <- balanceSamples(object = so.query, group = "Barcode")

}
\author{
Nicholas Mikolajewicz
}
