% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{benchmarkCluster}
\alias{benchmarkCluster}
\title{Examine cluster-specific scoring performance for cluster-specific gene sets that have been variably contaminated with random genes.}
\usage{
benchmarkCluster(
  object,
  benchmark_data,
  benchmark_genesets,
  which_cluster,
  miko.fdr.threshold = 0.05,
  coherence.fraction.threshold = 0
)
}
\arguments{
\item{object}{Seurat Object}

\item{benchmark_data}{`benchmark_data` returned in list from `benchmarkScores` function. Must run `benchmarkScores` prior to `benchmarkCluster`.}

\item{benchmark_genesets}{`benchmark_genesets` returned in list from `benchmarkScores` function. Must run `benchmarkScores` prior to `benchmarkCluster`.}

\item{miko.fdr.threshold}{FDR threshold for Miko scores. Default is 0.05.}

\item{coherence.fraction.threshold}{Coherence fraction threshold. Default is 0.8.}

\item{which.cluster}{Name of group within grouping variable used to run `benchmarkScores`. E.g., '3' if `benchmarkScores(... , group_by = "seurat_clusters")`}
}
\value{
list of benchmark results for specified cluster.
}
\description{
Examine cluster-specific scoring performance for cluster-specific gene sets that have been variably contaminated with random genes.
}
\seealso{
\code{\link{benchmarkScores}} for miko scoring benchmarking
}
\author{
Nicholas Mikolajewicz
}
