% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{benchmarkScores}
\alias{benchmarkScores}
\title{Benchmark Miko scoring pipeline using cluster-specific markers that had been variably contaminated with random genes.}
\usage{
benchmarkScores(
  object,
  geneset.size = 15,
  group_by = "seurat_clusters",
  assay = DefaultAssay(object),
  deg.logFC.threshold = 0.5,
  deg.fdr.threshold = 0.05,
  miko.fdr.threshold = 0.05,
  verbose = T,
  nworkers = 1
)
}
\arguments{
\item{object}{Seurat Object}

\item{geneset.size}{size of differentially-expressed gene sets}

\item{group_by}{Name of grouping variable in `object` meta feature. Default is "seurat_clusters".}

\item{assay}{Name of assay to use}

\item{deg.logFC.threshold}{logFC threshold for differential-expression analysis. Default is 0.5.}

\item{deg.fdr.threshold}{FDR threshold for differential-expression analysis. Default is 0.05.}

\item{miko.fdr.threshold}{FDR threshold for Miko scores. Default is 0.05.}

\item{verbose}{Print progress. Default is TRUE.}

\item{nworkers}{Number of workers for parallel implementation. Default is 1.}
}
\value{
list of benchmark results.
}
\description{
Differentially-expressed gene sets for each cluster are derived and variably contaminated with random genes prior to calculating cluster-specific miko scores for each gene set. The performance of miko scoring is then evaluated on cluster-specific and non-specific gene sets.
}
\seealso{
\code{\link{AddSModuleScore}} for standardized module scoring, \code{\link{wilcoxauc}} for differential expression analysis
}
\author{
Nicholas Mikolajewicz
}
\concept{miko_score}
