% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{binMatrix}
\alias{binMatrix}
\title{Row-wise matrix binning}
\usage{
binMatrix(m, bin.size)
}
\arguments{
\item{m}{matrix}

\item{bin.size}{Numeric indicating number of bins. Resulting matrix will have bin.size number of rows.}
}
\value{
binned matrix
}
\description{
Bin matrix row-wise by taking averages across regular row intervals.
}
\examples{

}
\author{
Nicholas Mikolajewicz
}
