% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{binVector}
\alias{binVector}
\title{Vector binning}
\usage{
binVector(v, bin.size)
}
\arguments{
\item{v}{numeric vector}

\item{bin.size}{Numeric indicating number of bins. Resulting vector will have bin.size length.}
}
\value{
binned vector
}
\description{
Bin vector by taking averages across regular intervals.
}
\examples{

}
\author{
Nicholas Mikolajewicz
}
