% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{categoricalColPal}
\alias{categoricalColPal}
\title{Generate categorical ColorBrewer palette.}
\usage{
categoricalColPal(labels = NULL, n = NULL, palette = "Spectral")
}
\arguments{
\item{labels}{vector of category names that if specified will have color assigned to. Default is NULL.}

\item{n}{number of colors. Default is NULL. Ignored if 'labels' is specified.}

\item{palette}{Name of palette. Default: "Spectral"}
}
\value{
Vector of colors. Named if labels provided.
}
\description{
Generate categorical ColorBrewer palette. Specify one of 'labels' or 'n' to generate categorical color palette for use with 'scale_fill_manual(values = col.pal)' and 'scale_color_manual(values = col.pal)'
}
\examples{

# generate color palette
col.pal <- categoricalColPal(n = 10)

# add to ggplot handle
gg.plt <- gg.plt + scale_fill_manual(values = col.pal)

}
\author{
Nicholas Mikolajewicz
}
