% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{checkGeneRep}
\alias{checkGeneRep}
\title{Check gene representation}
\usage{
checkGeneRep(reference.genes, query.genes)
}
\arguments{
\item{reference.genes}{Named vector of genes; names are ENSEMBL, entries are SYMBOL.}

\item{query.genes}{vector of query genes to check representation}
}
\value{
Character specifying ensembl or symbol
}
\description{
Check whether genes are represented as SYMBOL or ENSEMBL. This is accomplished by comparing query to reference, and using a majority vote.
}
\author{
Nicholas Mikolajewicz
}
