% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{citationCheck}
\alias{citationCheck}
\title{Check pubmed citations for genes}
\usage{
citationCheck(
  gene.query,
  search.query = NULL,
  delay = 0.5,
  mindate = 2000,
  maxdate = 2021,
  verbose = T,
  ...
)
}
\arguments{
\item{gene.query}{Genes to query. Character vector.}

\item{search.query}{Search query. Boolean expression are supported. Character vector. If search.query is not specified, the number of citations for each gene will be returned without constraining the tally to a specified topic.}

\item{delay}{Delay between queries. If database is queried too fast, error is thrown. Default delay is 0.5.}

\item{mindate}{Minimum of date range for search results (examples: 2002; 2002/01/01); must be supplied with maxdate.}

\item{maxdate}{Maximum of date range for search results; must be supplied with mindate.}

\item{verbose}{Print progress (T or F). Default is T.}

\item{...}{additional arguments passed to \code{\link{EUtilsSummary}}}
}
\description{
For given topic (search query), number of publications that mention specified genes (gene.query) are tallied and returned. Based on RISmed package by Stephanie Kovalchik.
}
\examples{
m8.citations <- citationCheck(gene.query = gene.list$m8, search.query = "hematopoietic|hematogenic|HSC", delay = 0.5,
mindate=2000, maxdate=2021, verbose = T)
}
\seealso{
\code{\link{EUtilsSummary}}
}
\author{
Nicholas Mikolajewicz
}
