% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cleanCluster}
\alias{cleanCluster}
\title{Clean clusters}
\usage{
cleanCluster(object, mad.threshold = 3, return.plots = F, verbose = T)
}
\arguments{
\item{object}{Seurat object.}

\item{mad.threshold}{Filter threshold. Represents minimal number of median absolute deviations from median required to omit cell from object. Default is 3.}

\item{return.plots}{Logical indicating whether to return plots. If true, returns list containing object along with ggplot handles. If false, only filtered object is returned.}

\item{verbose}{Logical indicating whether to print progress messages.}
}
\value{
list (return.plots == T) or seurat object (return.plots == F)
}
\description{
Filter cells that exceed N median absolute deviations from cluster center in UMAP coordinate space. Assumes that Seurat::FindClusters() has already been run.
}
\author{
Nicholas Mikolajewicz
}
