% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_functions.R
\name{clearGlobalEnv}
\alias{clearGlobalEnv}
\title{Remove variables from global environment}
\usage{
clearGlobalEnv(pattern, exact.match = T)
}
\arguments{
\item{pattern}{Character specifying name of variable(s) to remove from global environment.}

\item{exact.match}{If TRUE (default), exact pattern matching is enforced. If FALSE, partial pattern matching is facilitated with grep().}
}
\description{
Remove variables in global environment that match a given pattern.
}
\author{
Nicholas Mikolajewicz
}
