% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{coherentFraction}
\alias{coherentFraction}
\title{Calculate coherent fraction for feature expression program.}
\usage{
coherentFraction(
  object,
  score.matrix,
  genelist,
  method = c("pearson", "spearman"),
  assay = DefaultAssay(object),
  slot = "data",
  subsample.cluster.n = 500,
  nworkers = 1,
  verbose = T
)
}
\arguments{
\item{object}{Seurat Object.}

\item{score.matrix}{matrix of feature expression program. Can be computed using `AddModuleScores`, `AddSModuleScores`, `mikoScore`, among others.}

\item{method}{Correlation method: "pearson" or "spearman". Default is "pearson".}

\item{assay}{Name of assay to use.}

\item{slot}{Use expression data from this slot in `object`.}

\item{subsample.cluster.n}{number of cells to subsample within each cluster ("seurat_clusters" in `object` meta data). Default is 500.}

\item{nworkers}{Number of workers for parallel implementation. Default is 1.}

\item{verbose}{Print progress. Default is TRUE.}

\item{geneset}{gene set list used for obtaining `score.matrix` (e.g., gene set provided as input into `mikoScore` or `AddSModuleScores`).}
}
\value{
data.frame containing cluster-level coherent fractions.
}
\description{
Calculate fraction of genes that are correlated with feature expression program. Performed on a per-cluster basis ("seurat_clusters" in `object` meta data).
}
\examples{

so.query <-  AddSModuleScore(object = so.query, features = gene.list)

raw.mat <- so.query@misc[["raw_score"]]
colnames(raw.mat) <- gsub("raw_", "", colnames(raw.mat))
df.cscore <- coherentFraction(object = so.query, score.matrix =raw.mat, nworkers = 20,
                              genelist = gene.list, assay = DefaultAssay(so.query), slot = "data", subsample.cluster.n = 500)


}
\seealso{
\code{\link{AddSModuleScore}} for standardized module scoring, \code{\link{mikoScore}} for miko scoring
}
\author{
Nicholas Mikolajewicz
}
