% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{col2rowname}
\alias{col2rowname}
\title{Assign column entries in data.frame to row names.}
\usage{
col2rowname(df, col)
}
\arguments{
\item{df}{data.frame}

\item{col}{Character specifying column name in df.}
}
\value{
df
}
\description{
Assign column entries in data.frame to row names.
}
\examples{

f.mat <-aggGroupExpression(
so = so.query,
  which.data = "data",
  which.assay = DefaultAssay(so.query),
  which.center = "fraction",
  which.group = "seurat_clusters",
  do.parallel = F
)

f.mat <- col2rowname(f.mat, col = "genes")

}
\author{
Nicholas Mikolajewicz
}
