% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{consolidateNMF}
\alias{consolidateNMF}
\title{Consolidate several NMF reduction objects into single NMF reduction object within Seurat object.}
\usage{
consolidateNMF(object, reduction_key = "nmf")
}
\arguments{
\item{object}{Seurat object containing several NMF reductions (e.g., each NMF with different k parameter)}

\item{reduction_key}{Reduction key used to identify reduction objects for consolidation. Default is "nmf".}
}
\value{
Seurat object with with several NMF reductions consolidated into "nmf_all" reduction object.
}
\description{
Consolidate several NMF reduction objects into single NMF reduction object within Seurat object.
}
\references{
\url{https://nmikolajewicz.github.io/scMiko/articles/Module_Detection.html}
}
\seealso{
\code{\link{runNMF}}
}
\author{
Nicholas Mikolajewicz
}
