% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_functions.R
\name{da_DEG}
\alias{da_DEG}
\title{Identify correlated genes/pathways associated with differential abundance.}
\usage{
da_DEG(
  da.res.list,
  reduction = "pca",
  species = "Mm",
  red.gene.percentile = 0.99,
  do.enrich = F,
  pathway.db = "Bader"
)
}
\arguments{
\item{da.res.list}{da_Run output.}

\item{reduction}{Character specifying which reduction to use for differential abundance association analysis.}

\item{species}{Character specifying species. Either "Mm" or "Hs". Default is "Mm".}

\item{red.gene.percentile}{Numeric, top nth percentile genes/pathways.}

\item{do.enrich}{Logical to perform enrichment of DA-associated genesets.}

\item{pathway.db}{Character pathway database to use for enrichment. Ignored if do.enrich = F. Default: "Bader".}
}
\value{
updated da_Run output.
}
\description{
Identify correlated genes/pathways associated with differential abundance.
}
\author{
Nicholas Mikolajewicz
}
