% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_functions.R
\name{da_Run}
\alias{da_Run}
\title{Run differential abundance analysis.}
\usage{
da_Run(
  object,
  condition.group,
  sample.group = NA,
  design.groups = c(condition.group, sample.group),
  balance.samples = T,
  balance.size = NA,
  assay = DefaultAssay(object),
  reference.group = NULL,
  fdr.correction = F
)
}
\arguments{
\item{object}{Seurat object}

\item{condition.group}{Character specifying which metadata column to perform differential analysis on.}

\item{sample.group}{Character specifying metadata column with replicate groupings. If not specified (Default), pseudoreplicates are generated (see scMiko::pseudoReplicates()).}

\item{design.groups}{Character vector specifying model design. Default: c(condition.group, sample.group).}

\item{balance.samples}{Numeric specifying target sample size when balancing samples. 'balance.size' argument for scMiko::balanceSamples.}

\item{assay}{Character specifying which Seurat assay to use.}

\item{reference.group}{Character specifying which condition.group is reference.}

\item{fdr.correction}{Logical to perform FDR correction.}
}
\value{
list of results
}
\description{
Run differential abundance analysis. Adapted Milo method in a wrapper.
}
\author{
Nicholas Mikolajewicz
}
