% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{discretePalette}
\alias{discretePalette}
\title{Apply discrete color palette to ggplot object}
\usage{
discretePalette(gg, fc, n.groups)
}
\arguments{
\item{gg}{ggplot handle}

\item{fc}{character specifying whether to apply color palette to 'fill', 'color', or 'both'}

\item{n.groups}{numeric specifying number of groups}
}
\value{
ggplot2 handle
}
\description{
Apply discrete color palette to ggplot object
}
\examples{

gg.plot <- discretePalette(gg.plot, fc = "color", n.groups = 4)

}
