% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dist2hclust}
\alias{dist2hclust}
\title{Hierarchially-cluster distance matrix}
\usage{
dist2hclust(d.mat, method = "average", ...)
}
\arguments{
\item{d.mat}{distance matrix}

\item{method}{hierarchial-clustering method. Default is "average". See flashClust for options.}

\item{...}{Additional arguments passed to flashClust {flashClust package}}
}
\value{
hclust object
}
\description{
Hierarchially-cluster distance matrix using flashClust package.
}
\examples{

# Get expression matrix
which.data <- "scale"

# variable gene only matrix
use.var <- T
if (use.var){
  exp.mat <- getExpressionMatrix(so.query, only.variable = use.var, which.data = which.data, use.additional.genes = NA)
} else {
  exp.mat <- exp.mat.complete
}


# transpose expressio matrix (genes are columns)
t.exp.mat <- t(exp.mat)
datExpr <- as.matrix(t.exp.mat)
SubGeneNames=colnames(datExpr)

# capture output used to hide undesired print statement
print2hide <- capture.output(allowWGCNAThreads())

# transform matrix if necessary
if (min(datExpr) < 0) {
  datExpr.noz <- datExpr + abs(min(datExpr))
} else {
  datExpr.noz <- datExpr
}

# run WGCNA
output.all <- runWGCNA(datExpr.noz, cor.metric = "rho_p", soft.power = 2, use.TOM = T)

# unpack output
s.mat <- output.all[["s.mat"]] # similar matrix
a.mat <- output.all[["a.mat"]] # adjacency matrix
w.mat <- output.all[["w.mat"]] # topological overlap matix
d.mat <- output.all[["d.mat"]] # disimilarity matix

geneTree <- dist2hclust(d.mat)

}
\seealso{
\code{\link{flashClust}}
}
\author{
Nicholas Mikolajewicz
}
