% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{downsampleSeurat}
\alias{downsampleSeurat}
\title{Downsample single cell data}
\usage{
downsampleSeurat(object, subsample.factor = 1, subsample.n = NULL, verbose = T)
}
\arguments{
\item{object}{Seurat Object}

\item{subsample.factor}{Numeric [0,1]. Factor to downsample data by.}

\item{subsample.n}{Numeric [1,ncol(object)]. Number of cells to subsample. If specified, overides subsample.factor.}

\item{verbose}{Print progress. Default is TRUE.}
}
\value{
Seurat Object
}
\description{
Downsample number of cells in Seurat object by specified factor
}
\author{
Nicholas Mikolajewicz
}
