% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dpt_for_branch}
\alias{dpt_for_branch}
\title{Get branch-specific pseudotimes from diffusion map}
\usage{
dpt_for_branch(dpt, branch_id)
}
\value{
numeric vector; pseudotimes
}
\description{
Get branch-specific pseudotimes from diffusion map. Adopted from destiny package.
}
\examples{

# PCA embedding
so.pca <- so.query@reductions[["pca"]]@cell.embeddings

# diffusion map
dm <- DiffusionMap(so.pca, n_eigs = 50)

# diffusion pseudotime
dpt <- DPT(dm, tips =tip.cluster)

# flatten
dpt_flat <- branch_divide(dpt, divide = 1)

# get branch specific pseudotime
pt_vec <- dpt_for_branch(dpt_flat, branch_id = paths_to)

}
\seealso{
\code{\link{DPT}}
}
