% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{enrichGO.fisher}
\alias{enrichGO.fisher}
\title{Over-representation enrichment of GO terms using weighted fisher method}
\usage{
enrichGO.fisher(
  gene.candidates,
  gene.universe,
  which.species,
  which.ontology = "BP"
)
}
\arguments{
\item{gene.candidates}{Vector of candidate genes.}

\item{gene.universe}{Vector of all genes.}

\item{which.species}{Character specifying species. "Mm" or "Hs".}

\item{which.ontology}{Character specifying ontology. "BP", "MF", or "CC".}
}
\value{
list of 2 data.frames: unadjusted.results (p<0.001) and adjusted.results (padj < 0.05)
}
\description{
Over-representation enrichment of GO terms using weighted fisher method
}
