% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ens2sym.so}
\alias{ens2sym.so}
\title{Convert gene ensemble to symbol in seurat object}
\usage{
ens2sym.so(object, gNames.list, convert.RNA = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{gNames.list}{named character vector mapping ensembl (names) to gene symbols (entries)}

\item{convert.RNA}{logical to convert ensembl in RNA assay to gene symbol}
}
\value{
Seurat object with genes represented as symbols
}
\description{
Process Seurat object so that genes in Ensemble format are converted to Symbol representation.
}
\author{
Nicholas Mikolajewicz
}
