% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{entrez2sym}
\alias{entrez2sym}
\title{Convert entrez id to gene symbol}
\usage{
entrez2sym(my.entrez, my.species)
}
\arguments{
\item{my.species}{Character. Species.}

\item{my.symbols}{Character. Vector of Entrez IDs.}
}
\value{
data.frame mapping Entrez IDs to gene Symbols.
}
\description{
Entrez ID is converted to gene symbol using org.Hs.eg.db or org.Mm.eg.db annotation databases.
}
