% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{exprUMAP}
\alias{exprUMAP}
\title{Visualize gene expression on UMAP}
\usage{
exprUMAP(object, feature, plot.subtitle = NULL, do.neb = T, title.size = 10)
}
\arguments{
\item{object}{Seurat object.}

\item{feature}{feature name.}

\item{plot.subtitle}{Plot title.}

\item{do.neb}{Logical to use Nebulosa::plot_density instead of scMiko::scExpression.UMAP. Default is T.}

\item{title.size}{Size of plot title. Default is 10.}
}
\value{
ggplot handle
}
\description{
Visualize gene expression on UMAP. Wrapper for Nebulosa::plot_density() and scMiko::scExpression.UMAP() functions.
}
\examples{

 gg.plot <-  exprUMAP(object = so, feature = "Prrx1")

}
