% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{featureGradient}
\alias{featureGradient}
\title{Visualize feature activity/expression gradient overlaid on UMAP}
\usage{
featureGradient(
  object,
  feature,
  umap.key = "umap",
  min.quantile.cutoff = 0.025,
  max.quantile.cutoff = 0.975,
  arrow.color = "black",
  arrow.size = 1
)
}
\arguments{
\item{object}{Seurat object.}

\item{feature}{scaling constant used to determine optimal point size.}

\item{umap.key}{name of UMAP reduction. Default is "umap". Requires that UMAP embedding was calculated using Seurat::RunUMAP() prior to running scMiko::featureGradient().}

\item{min.quantile.cutoff}{upper quantile threshold at which values are winsorized. Default is 0.975}

\item{max.quantile.cutoff}{lower quantile threshold at which values are winsorized. Default is 0.025}

\item{arrow.color}{gradient arrow color. Default is "black".}

\item{arrow.size}{gradient arrow size Default is 1.}
}
\value{
ggplot handle
}
\description{
Visualize feature activity/expression gradient overlaid on UMAP
}
\examples{

 gg.plot <-  featureGradient(object = so, feature = "S.Score")

}
