% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_functions.R
\name{findCDIMarkers}
\alias{findCDIMarkers}
\title{Calculate feature dependency index}
\usage{
findCDIMarkers(
  object,
  features.x = NULL,
  features.y = rownames(object),
  ncell.subset = 5000,
  assay = DefaultAssay(object),
  slot = "data",
  verbose = T
)
}
\arguments{
\item{object}{Seurat object}

\item{features.x}{feature or meta feature. CDI between features.x and features.y are computed.}

\item{features.y}{feature or meta feature. CDI between features.x and features.y are computed.}

\item{ncell.subset}{max number of cells to run analysis on. Default is 5000. Computationally intensive for larger datasets.}

\item{assay}{Assay to run CDI scoring on. Default is DefaultAssay(object).}

\item{slot}{slot to run CDI scoring on. Default is data.}

\item{verbose}{print progress. Default is T.}
}
\value{
data.frame with CDI scores.
}
\description{
Calculate feature dependency index (CDI).
}
\examples{

}
\seealso{
\code{\link{binom.test}}
}
\author{
Nicholas Mikolajewicz
}
