% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_functions.R
\name{FindConservedCDIMarkers}
\alias{FindConservedCDIMarkers}
\title{Find CDI-derived markers that are conserved across samples}
\usage{
FindConservedCDIMarkers(
  object,
  features.x,
  features.y = rownames(object),
  group.by,
  n.workers = 1,
  verbose = T,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{features.x}{feature or meta feature. CDI between features.x and features.y are computed.}

\item{features.y}{feature or meta feature. CDI between features.x and features.y are computed.}

\item{group.by}{meta feature used to group data into independent samples.}

\item{n.workers}{number of workers for parallel implementation. Default is 1 (no parallel).}

\item{verbose}{print progress. Default is T.}

\item{...}{additional arguments passed to findCDIMarkers(...)}
}
\value{
data.frame with pooled CDI scores.
}
\description{
Find CDI-derived markers that are conserved across samples. For each sample (i.e., specified using group.by argument), FindCDIMarkers() is run, and the outputs from each sample are pooled. CDI scores are aggregated as means, whereas p values are pooled using the Fisher Method.
}
\examples{

cdi.conserved <- FindConservedCDIMarkers(object = so.ps, features.x = "celltype", group.by = "sample", n.workers = 15)

}
\seealso{
\code{\link{findCDIMarkers}}
}
\author{
Nicholas Mikolajewicz
}
