% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_functions.R
\name{findCorMarkers}
\alias{findCorMarkers}
\title{Calculate spearman correlations between features.}
\usage{
findCorMarkers(
  object,
  features.x = NULL,
  features.y = rownames(object),
  ncell.subset = 5000,
  geosketch.subset = F,
  assay = DefaultAssay(object),
  slot = "data",
  verbose = T
)
}
\arguments{
\item{object}{Seurat object}

\item{features.x}{feature or meta feature. Spearman correlation between features.x and features.y are computed.}

\item{features.y}{feature or meta feature. Spearman correlation between features.x and features.y are computed.}

\item{ncell.subset}{max number of cells to run analysis on. Default is 5000.}

\item{geosketch.subset}{Use GeoSketch method to subsample scRNA-seq data while preserving rare cell states (https://doi.org/10.1016/j.cels.2019.05.003). Logical, T or F (Default F). Recommended if cell type representation is imbalanced.}

\item{assay}{Assay to run spearman correlation on. Default is DefaultAssay(object).}

\item{slot}{slot to run spearman correlation on. Default is data.}

\item{verbose}{print progress. Default is T.}
}
\value{
data.frame with spearman correlations.
}
\description{
Calculate spearman correlations between features in Seruat object. Sparse implementation enables faster calculation of spearman correlations without need to cast sparse expression matrix to dense matrix.
}
\author{
Nicholas Mikolajewicz and Saket Choudhary (https://github.com/saketkc/blog/blob/main/2022-03-09/SparseSpearmanCorrelation.ipynb)
}
