% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_functions.R
\name{findGiniMarkers}
\alias{findGiniMarkers}
\title{Calculate Gini marker specificity}
\usage{
findGiniMarkers(
  object,
  features = NULL,
  group.by = "seurat_clusters",
  assay = NULL,
  min.pct = 0.1,
  verbose = T
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{feature to run gini specificy scoring on.}

\item{group.by}{groups used for feature specificity assignment. E.g., Each gene is assigned to a single group in which it exhibits maximal specificity. Default is 'seurat_clusters'.}

\item{assay}{Assay to run  gini specificity scoring on. Default is DefaultAssay(object).}

\item{min.pct}{minimal expression fraction per group. Gini scoring performs poorly on genes expressed at extremely low levels. Default (recommended) is 0.1.}

\item{verbose}{print progress. Default is T.}
}
\value{
data.frame with feature-specific gini scores.
}
\description{
Calculate Gini marker specificity.
}
\examples{

}
\seealso{
\code{\link{fdrtool}} for p values calculations, \code{\link{ineq}} for gini index calculations.
}
\author{
Nicholas Mikolajewicz
}
