% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fixBarcodeLabel}
\alias{fixBarcodeLabel}
\title{Fix barcode labels}
\usage{
fixBarcodeLabel(so)
}
\arguments{
\item{so}{Seurat Object}
}
\value{
Seurat object
}
\description{
Rename metadata entry 'CellTypes' to 'Barcode'. This is a fix implemented to correct an error from an earlier analysis pipeline. Analyses post January 2020 do not require this fix.
}
\author{
Nicholas Mikolajewicz
}
