% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_functions.R
\name{flex.asDT}
\alias{flex.asDT}
\title{Outputs datatable with print buttom options}
\usage{
flex.asDT(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.table
}
\description{
Outputs datatable with print buttom options
}
