% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_functions.R
\name{flex.multiTabLogs}
\alias{flex.multiTabLogs}
\title{Generate multi-tab analysis log list for flexdashboard}
\usage{
flex.multiTabLogs(module.logs)
}
\arguments{
\item{module.logs}{Character vector specifying name of prior analysis logs available in global enviroment. Must have "df.log_Module_" stem.}
}
\value{
flexdashboard compatable list of analysis logs
}
\description{
Prepares list of prior analysis logs for multi-tab table presentation in flexdashboards.
}
\examples{

'''{r multi tab log}
out <- flex.multiTabLogs(module.logs)
'''

`r paste(knitr::knit(text = paste(out, collapse = '\n')))`

}
