% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_functions.R
\name{flex.multiTabTables}
\alias{flex.multiTabTables}
\title{Generate multi-tab datatable list for flexdashboard}
\usage{
flex.multiTabTables(df.list, df.list.name)
}
\arguments{
\item{df.list}{Named list of data.frames.}

\item{df.list.name}{Character specifying name of df.list.}
}
\value{
flexdashboard compatable list of tables.
}
\description{
Prepares list of data.frames for multi-tab table presentation in flexdashboards.
}
\examples{
\dontrun{
Tab Header 1
=====================================

Row {.tabset}
-------------------------------------

'''{r multi tab table}
out <- flex.multiTabTables(df.list, "df.list")
'''

`r paste(knitr::knit(text = paste(out, collapse = '\n')))`
}

}
