% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{generateHeat}
\alias{generateHeat}
\title{Generate heatmap with plotly interface}
\usage{
generateHeat(
  mat,
  heat.object,
  scale.limit = NULL,
  legend.label = "",
  x.label = "",
  y.label = "",
  title = ""
)
}
\arguments{
\item{mat}{matrix. Must be same matrix provided to scMiko::getHeat().}

\item{heat.object}{output from scMiko::getHeat()}

\item{scale.limit}{Numeric specifying heatmap limits. If unspecified, defined as max(c(abs(max(mat)), abs(min(mat)))).}

\item{legend.label}{Character.}

\item{x.label}{x axis label. Character}

\item{y.label}{y axis label. Character.}

\item{title}{plot title. Character.}
}
\value{
plotly handle
}
\description{
Generate heatmap with plotly interface. Takes output from scMiko::getHeat() function as input.
}
