% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{geneRepCurve}
\alias{geneRepCurve}
\title{Plot relationship showing percentage of cells expressing atleast % of genes}
\usage{
geneRepCurve(so, which.genes = NULL, only.variable = F, which.data = "data")
}
\arguments{
\item{so}{Seurat object}

\item{which.genes}{Specify geneset. If null, all gene are included (unless only.variable is set to True)}

\item{only.variable}{Only used variable genes}

\item{which.data}{specify which data slot to use (scale or data; data is default)}
}
\value{
ggplot handle
}
\description{
Plot relationship showing percentage of cells expressing atleast % of genes
}
