% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{geneSets}
\alias{geneSets}
\title{List of scMiko gene sets}
\format{

}
\usage{
data(geneSets)
}
\description{
Reference gene sets consolidated from numerous sources to faciliate cell annotation and module scoring.
}
\examples{
data(geneSets)
cell_catalog <- geneSets[["Cell_Catalog"]]

}
\keyword{annotation,}
\keyword{cell}
\keyword{genesets,}
\keyword{module}
\keyword{scoring}
