% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{Split violin plot using ggplot2}
\usage{
geom_split_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "identity",
  ...,
  draw_quantiles = NULL,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{ggplot mapping}

\item{data}{data.frame}

\item{stat}{ggplot identity}

\item{position}{ggplot position}

\item{...}{additional parameters to pass to layer()}

\item{draw_quantiles}{draw quantiles. Default is NULL}

\item{trim}{Trim data. Default is True.}

\item{scale}{Default is area.}

\item{na.rm}{Default is na.rm}

\item{show.legend}{Default is NA}

\item{inherent.aes}{Default is TRUE}
}
\value{
ggplot object
}
\description{
Draws split violin plot using ggplot2. Credit to jan-glx (https://stackoverflow.com/questions/35717353/split-violin-plot-with-ggplot2)
}
\examples{

# split violin plot
ggplot(my_data, aes(x, y, fill = m)) + geom_split_violin()


}
