% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getAvailableFiles}
\alias{getAvailableFiles}
\title{Get list of available files}
\usage{
getAvailableFiles(directory = "Preprocessed Datasets/")
}
\arguments{
\item{directory}{Character. Directory name.}

\item{scale.limit}{Numeric. Color scale limit.}
}
\value{
list of files
}
\description{
Get list of files in specified directory. Default directory is "Preprocessed Datasets/".
}
