% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getClusterCenters}
\alias{getClusterCenters}
\title{Get cluster centers}
\usage{
getClusterCenters(df, which.center = "mean")
}
\arguments{
\item{df}{data.frame with columns 'x' and 'y' specifying dimensional reduction coordinates, and column 'cluster' or 'clusters' specifying cluster membership.}

\item{which.mean}{Specify which central values to use. One of 'mean' or 'median'. Default is 'mean'.}
}
\value{
df.centers
}
\description{
For given dimensional reduction (e.g., UMAP), return cluster center coordinates.
}
\examples{

# compute cluster centers
which.center <- "mean"
df.red.centers <- get.cluster.centers(df.red, which.center)

}
