% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getClusterRoot}
\alias{getClusterRoot}
\title{Get cell index corresponding to center of specified root cluster.}
\usage{
getClusterRoot(x, y, cluster.membership, which.cluster)
}
\arguments{
\item{x}{x coordinates (e.g., UMAP 1)}

\item{y}{y coordinates (e.g., UMAP 2)}

\item{cluster.membership}{group id's corrorespodning to cluster membership (e.g., seurat_cluster entries). Must be same length as x and y.}

\item{which.cluster}{specify which cluster to get root cell fot.}
}
\value{
root cell index
}
\description{
Get cell index corresponding to center of specified root cluster. For specified cluster, cluster center is computed and index of nearest cell to cluster center is returned.
}
\examples{

}
