% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getConnectivity}
\alias{getConnectivity}
\title{Gene connectivity within network.}
\usage{
getConnectivity(w.mat, gene.names, flag.top.n = 20)
}
\arguments{
\item{w.mat}{Adjaceny or toplogoical overlap matrix. Symmetrical matrix, with row and column-wise gene entries.}

\item{gene.names}{Vector of gene names corresponding to genes in coloumns/rows of w.mat.}

\item{flag.top.n}{Numerical specifying top N genes to flag (for subsequent plotting). Default is 20.}
}
\value{
Dataframe with gene connectivity, rank and label flag.
}
\description{
Computes gene connectivity from adjaceny/topological overlap matix. Part of WCGNA network analysis workflow.
}
\examples{

# run WCGNA
output.all <- runWCGNA(datExpr.noz, cor.metric = "rho_p", soft.power = 2, use.TOM = T)

# unpack output
s.mat <- output.all[["s.mat"]] # similar matrix
a.mat <- output.all[["a.mat"]] # adjacency matrix
w.mat <- output.all[["w.mat"]] # topological overlap matix
d.mat <- output.all[["d.mat"]] # disimilarity matix

# get connectivity
df.con <- getConnectivity(w.mat, gene.names = colnames(a.mat))

}
\author{
Nicholas Mikolajewicz
}
