% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getDensity}
\alias{getDensity}
\title{Get local density (z) of bivariate relationship (x,y)}
\usage{
getDensity(x, y, ...)
}
\arguments{
\item{x}{numeric (equal length as y)}

\item{y}{numeric}

\item{...}{additional parameters passed to MASS::kde2d}
}
\value{
density values
}
\description{
Get local density (z) of bivariate relationship (x,y)
}
\examples{
# get data and compute densities
df.meta <- so@meta.data # so is seurat object
df.meta$density1 <- getDensity(df.meta$nCount_RNA, df.meta$percent.mt, n = 100)

# generate scatter plot with overlayed density values
plt.handle1 <- df.meta \%>\% ggplot(aes(x = nCount_RNA, y = percent.mt, color = density1)) + geom_point() + theme_miko(legend = T) +
   xlab("UMI/cell") + ylab("Mitochondrial Content (\%)") +
   labs(title = paste0("r = ", rho1p, "; rho = ", rho1s)) + scale_color_viridis("Density") +
   theme(legend.position="bottom", legend.key.width=unit(legend.width,"cm"))

}
\author{
Nicholas Mikolajewicz

Kamil Slowikowski (https://slowkow.com/notes/ggplot2-color-by-density/)
}
