% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getExpressingCells}
\alias{getExpressingCells}
\title{Get cells that express query gene}
\usage{
getExpressingCells(so, query, expression.threshold = 0, which.data = "data")
}
\arguments{
\item{so}{Seurat Object}

\item{query}{Query gene. A character.}

\item{expression.threshold}{Numeric. Return cells that express query above threshold value. Default is 0.}

\item{which.data}{Seurat data slot ("data", "scale")}
}
\value{
Character vector of cell barcodes/ids
}
\description{
Get cells that express query gene above specified threshold
}
\author{
Nicholas Mikolajewicz
}
