% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getExpressionMatrix}
\alias{getExpressionMatrix}
\title{Get scaled expression matrix from Seurat Object}
\usage{
getExpressionMatrix(so, only.variable = F, which.assay = NULL)
}
\arguments{
\item{so}{Seurat Object}

\item{only.variable}{Logical indicating whether to include variable features only or not.}

\item{which.assay}{Seurat assay to get data frome. Default is DefaultAssay(so).}
}
\value{
gene x cell expression matrix
}
\description{
Get scaled expression matrix from Seurat Object
}
