% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getHeat}
\alias{getHeat}
\title{Create heatmap object}
\usage{
getHeat(mat, hmcol, scale.limit, ...)
}
\arguments{
\item{mat}{Matrix. Input matrix for heatmap.}

\item{hmcol}{Heatmap colors}

\item{scale.limit}{Numeric. Color scale limit.}

\item{...}{additional parameters passed to gplots::heatmap.2(...)}
}
\value{
Heatmap object
}
\description{
Uses gplots::heatmap.2 function to generate heatmap object. Run in chunk specifying include = F to suppress plot and only output object.
}
