% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{getICAGenes}
\alias{getICAGenes}
\title{Get significant ICA genes}
\usage{
getICAGenes(
  feature.loading,
  fdr.cutoff = 1e-04,
  local.fdr = T,
  assert.positive.skew = T,
  only.pos = T
)
}
\arguments{
\item{feature.loading}{ICA feature loadings}

\item{fdr.cutoff}{FDR threshold. Default is 0.0001.}

\item{local.fdr}{Whether to use local FDR. Default is T.}

\item{assert.positive.skew}{Whether to ensure that max(|feature.loading|) is positive. Default is T.}

\item{only.pos}{Whether significant genes are positively loaded only. Default is T.}
}
\value{
named list of ICA gene programs
}
\description{
Given feature loadings obtained from independent component analysis (ICA), significant genes for each component are identified.
}
\examples{

}
\references{
\url{https://nmikolajewicz.github.io/scMiko/articles/Module_Detection.html}
}
\seealso{
\code{\link{runICA}} for independent component analysis.
}
\author{
Nicholas Mikolajewicz
}
