% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getJaccard}
\alias{getJaccard}
\title{Jaccard Similarity}
\usage{
getJaccard(x1, x2)
}
\arguments{
\item{x1}{set 1}

\item{x2}{set 2}
}
\value{
Jaccard similarity score (numeric)
}
\description{
Computes Jaccard Similarity between two sets, x1 and x2.
}
\examples{

# compute jaccard similarity matrix for (named) list of genesets.
n.sets <- length(gene.sets)
j.mat <- matrix(nrow = n.sets, ncol = n.sets)
for (i in 1:n.sets){
  for (j in 1:n.sets){
    i.name <- names(gene.sets)[i]
    j.name <- names(gene.sets)[j]
    j.mat[i, j] <- getJaccard(gene.sets[[i.name]], gene.sets[[j.name]])
  }
}

# assign row and col names
rownames(j.mat) <- names(gene.sets)
colnames(j.mat) <- names(gene.sets)

# generate heatmap
pheatmap::pheatmap(j.mat, show_colnames = F, main = "Jaccard Similarity")

}
