% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_functions.R
\name{getLoadPath}
\alias{getLoadPath}
\title{Return load path}
\usage{
getLoadPath(file, directory = NULL)
}
\arguments{
\item{file}{file name}

\item{directory}{directory}
}
\value{
Character. Load path.
}
\description{
For specified file and directory, return a load path
}
