% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_functions.R
\name{getMissingPackages}
\alias{getMissingPackages}
\title{install missing packages from cran or bioconductor}
\usage{
getMissingPackages(
  package.list,
  install.missing = F,
  prefer.repo = "bioconductor"
)
}
\arguments{
\item{package.list}{Character vector of package names.}

\item{install.missing}{Flag specfiying whether to install missing packages. If FALSE, return list of missing packages only. If none are missing, NULL is returned.}

\item{prefer.repo}{Preffered repository for installing packages. If package is available on both repos, prefferred is use. Options are 'bioconductor' (default) or 'cran'.}
}
\description{
Install missing packages from cran or bioconductor, and flag any remaining packages that are not available on either repository.
}
\author{
Nicholas Mikolajewicz
}
