% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_functions.R
\name{getMitoContent}
\alias{getMitoContent}
\title{Calculate mitochondrial content}
\usage{
getMitoContent(so, gNames = NULL, assay = NULL, omit.na = T)
}
\arguments{
\item{so}{Seurat Object}

\item{gNames}{Named gene vector; entries are gene symbols, names are Ensembl.}

\item{assay}{assay to use.}

\item{omit.na}{Logical specifying whether to omit NA entries (present when unfiltered 10x dataset is used). Default is True.}
}
\value{
Seurat Object (with updated metadata)
}
\description{
Calculate cell-level mitochondrial content based on mapped transcripts and store in Seurat "percent.mt' metadata field.
}
