% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getModuleGenes}
\alias{getModuleGenes}
\title{Get list of module genes}
\usage{
getModuleGenes(module.colors, genes, add.prefix = T)
}
\arguments{
\item{module.colors}{Vector of module membership (usually colors). Must be same length as gene list.}

\item{genes}{Vector of genes.}

\item{add.prefix}{Logical specifying whether to add "M#." prefix to each module name.}
}
\value{
Named list, where names are module names and entries are module genes.
}
\description{
Map module membership to gene list and return module gene list.
}
\examples{

module.list.1 <- getModuleGenes(modules.1, SubGeneNames, add.prefix = F)
module.list.2 <- getModuleGenes(modules.2_new, SubGeneNames, add.prefix = F)

# get genes that are common to both network modules
module.union.list <- list()
common.modules <- intersect(modules.1, modules.2_new)
for (i in 1:length(common.modules)){
  common.genes <- intersect(module.list.1[[common.modules[i]]], module.list.2[[common.modules[i]]])
  module.union.list[[common.modules[i]]] <- common.genes
}


}
