% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getModulePreservation}
\alias{getModulePreservation}
\title{Determine module preservation between reference and query network}
\usage{
getModulePreservation(
  ref.mat,
  query.mat,
  ref.module,
  networkType = "unsigned",
  ...
)
}
\arguments{
\item{ref.mat}{reference data (expression matrix, cols are genes and rows and samples)}

\item{query.mat}{query data (expression matrix, cols are genes and rows and samples)}

\item{ref.module}{reference module membership. Vector of colors (length is equal to number of samples), specfiying sample membership to each module.}

\item{networkType}{Allowed values are (unique abbreviations of) "unsigned", "signed", "signed hybrid". See WGCNA::adjacency.}

\item{...}{Additional arguments passessed to modulePreservation {WGCNA package}}
}
\value{
data.frame of module preservation statistics
}
\description{
Determine module preservation between reference and query network using expression matrices from two scRNAseq comparison groups. Uses WGCNA::modulePreservation() to assess how well a module in one sample is preserved in another. 5<Z<10 indicates moderate presevation, while Z>10 indicates high preservation. Grey module contains uncharacterized genes while gold module contains random genes (these are used as controls).
}
\examples{

mColorh.1 <- optimalDS(dist2hclust(d.1), d.1, pamStage = F,cutHeight = 0.99)
modules.1 = mColorh.1[,4]

stats <- getModulePreservation(ref.mat = de.1, query.mat = de.2, modules.1, nPermutations=30,maxGoldModuleSize=100,maxModuleSize=400, verbose=3)

}
