% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getNMFGenes}
\alias{getNMFGenes}
\title{Returns top module genes from NMF feature loading matrix}
\usage{
getNMFGenes(feature.loading, norm.cutoff = 0.5)
}
\arguments{
\item{feature.loading}{Matrix; m genes x n modules loading}

\item{norm.cutoff}{Numeric [0,1]; cutoff to get top genes. Default is 0.5.}
}
\description{
Returns top module genes from NMF feature loading matrix
}
\examples{

# get feature loading matrix (W matrix)
nmf.kme <- as.matrix(so.query@misc[["nmf"]][["nmf"]][["W"]])

# get module genes
nmf.module.genes <- getNMFGenes((nmf.kme), norm.cutoff = (1/ncol(nmf.kme)) + .25)

}
\author{
Nicholas Mikolajewicz
}
