% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getOrderedGroups}
\alias{getOrderedGroups}
\title{Get vector of unique ordered group names from Seurat Object}
\usage{
getOrderedGroups(so, which.group = "seurat_clusters", is.number = T)
}
\arguments{
\item{so}{Seurat Object}

\item{which.group}{Character specfying group field in Seurat metadata. Default is "seurat_clusters".}

\item{is.number}{Logical specifying whether groups IDs are numeric. Default is True.}
}
\value{
Ordered vector
}
\description{
Get vector of unique ordered group names from Seurat Object
}
