% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getReductionGenes}
\alias{getReductionGenes}
\title{Get top loaded features for PCA or ICA dimensional reduction.}
\usage{
getReductionGenes(feature.loading, sig.threshold = 0.99, top.is.positive = T)
}
\arguments{
\item{feature.loading}{Feature loading matrix}

\item{sig.threshold}{Numeric specifying top nth percentile of features to return. Default: 0.99.}

\item{top.is.positive}{Logical specifying whether to enforce top loaded feature being positive. If negative, loading sign is inverted.}
}
\value{
list of ggplot handles
}
\description{
Get top loaded features for PCA or ICA dimensional reduction.
}
\examples{

feature.loading <- as.matrix(seurat.object@reductions[[reduction]]@feature.loadings)
red.res <- getReductionGenes(feature.loading = feature.loading,  sig.threshold = 0.99)

}
\author{
Nicholas Mikolajewicz
}
