% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getSoftThreshold.v2}
\alias{getSoftThreshold.v2}
\title{Analysis of scale free topology for soft-threshold. Modified from getSoftThreshold.}
\usage{
getSoftThreshold.v2(
  s.mat,
  power = c(seq(0.5, 5, by = 0.5), seq(6, 10)),
  network.type = "signed",
  nBreaks = 20,
  removeFirst = T,
  rescale.adjacency = F
)
}
\arguments{
\item{s.mat}{similarity matrix}

\item{power}{Numeric vector of powers to evaluate. Default is c(seq(0.5,5, by = 0.5), seq(6,10))}

\item{nBreaks}{Number of bins in connectivity histograms. Default is 20.}

\item{removeFirst}{Logical specifying whether the first bin should be removed from the connectivity histogram. Default is True.}

\item{rescale.adjacency}{Logical indicating if s.mat should be rescaled to [0,1]}

\item{networkType}{Allowed values are (unique abbreviations of) "unsigned", "signed", "signed hybrid". See WGCNA::adjacency.}
}
\value{
named list containing power estimates, r2 estimates, distribution plots, optimiation plot and results data.frame.
}
\description{
Analysis of scale free topology for multiple soft thresholding powers. The aim is to help the user pick an appropriate soft-thresholding power for network construction. Inspired by WGCNA::pickSoftThreshold and updated from first version (scMiko::getSoftThreshold)
}
\examples{

# determine optimal soft threshold
sft <- getSoftThreshold.v2(s.mat, power =c(seq(0.5,5, by = 0.5), seq(6,10)),
network.type = "signed", rescale.adjacency = F)

# visualize optimization plot
print(sft$optimization.plot)

# visualize node-linkage density plots
cowplot::plot_grid(plotlist = sft$distribution.plot, ncol = 5)
}
