% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{group2list}
\alias{group2list}
\alias{pruneSSN}
\title{Named list of cells grouped by meta feature from Seurat object}
\usage{
group2list(object, group = "seurat_clusters", is.num = F, prefix = "")

pruneSSN(object, graph = "RNA_snn_power", prune.threshold = 0.1, return.df = F)
}
\arguments{
\item{object}{Seurat object objected from SSN analysis.}

\item{group}{grouping feature (must be present in `object` meta data). Default is "seurat_clusters".}

\item{is.num}{whether `group` is a numerical feature.}

\item{prefix}{prefix added to each named entry in list. Default is "".}

\item{graph}{name of shared nearest neighbor (SNN) graph used for SSN analysis.}

\item{prune.threshold}{threshold used for pruning features from SSN graph (features with low connectivity are preferentially pruned). Default is 0.1.}

\item{return.df}{whether to return data.frame instead of named list.}
}
\value{
Returns named list of cells grouped by meta feature from Seurat object.

Returns named list of SSN module features.
}
\description{
Returns named list of cells grouped by meta feature from Seurat object.

Identify and (optionally) prune module features in scale-free shared nearest neighbor network (SSN)
}
\examples{


}
\seealso{
\code{\link{runSSN}} for SSN analysis
}
\author{
Nicholas Mikolajewicz

Nicholas Mikolajewicz
}
