% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{highlightUMAP}
\alias{highlightUMAP}
\title{Highlight cells on UMAP plot}
\usage{
highlightUMAP(
  object,
  group = "seurat_clusters",
  reduction = "umap",
  highlight.color = "tomato"
)
}
\arguments{
\item{object}{Seurat object.}

\item{group}{grouping variable in object's meta data.}

\item{reduction}{reduction used to visualize cells. Default is "umap".}

\item{highlight.color}{color of highlighted cells. Default is "tomato".}
}
\value{
list of ggplot handles
}
\description{
Highlight cells on UMAP plot
}
\examples{

 plot.highlight <-  highlightUMAP(object = so, group = "seurat_clusters")

}
