% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{id2geneset}
\alias{id2geneset}
\title{Get Reactome/GO geneset}
\usage{
id2geneset(id, my.species)
}
\arguments{
\item{id}{Reactome/GO identifier. Reactome has "R-" prefix; GO has "GO" prefix.}

\item{my.species}{Specify species; ensures that correct gene symbols are retrieved.}
}
\value{
Character vector of gene symbols belonging to Reactome/GO geneset.
}
\description{
Retrieves Reactome/GO geneset using specified annotation ID
}
\author{
Nicholas Mikolajewicz
}
