% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{id2term}
\alias{id2term}
\title{Map Reactome/GO ID to term}
\usage{
id2term(id, db = "Reactome", species = NULL)
}
\arguments{
\item{db}{Database to use for mapping. One of:
\itemize{
\item "Reactome" - Default
\item "GO"
}}

\item{species}{Character specifying species. If specified, one of:
\itemize{
\item "Hs"
\item "Mm"
}}

\item{term}{Reactome/GO ID}
}
\value{
Reactome/GO term
}
\description{
Map Reactome/GO ID to term
}
