% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{inferInitialTrajectory.v2}
\alias{inferInitialTrajectory.v2}
\title{Infer initial trajectory through space, using pseudotime priors}
\usage{
inferInitialTrajectory.v2(pt, space, w_width = 0.1)
}
\arguments{
\item{pt}{pseudotimes. Numeric vector.}

\item{space}{A numeric matrix or a data frame containing the coordinates of samples.}

\item{w.width}{the length of the smoothing window, if an integer, represents number of items, else, if a value between 0 and 1, represents the proportion of the input vector}
}
\value{
the initial trajectory
}
\description{
Infer initial trajectory through space, using pseudotime priors. Adopted from destiny package, uses smth.gaussian function; The specific function for smoothing using the gaussian window function.
}
\examples{

 # lineage-specific pseudotime
 pt_vec <- dpt_for_branch(dpt_flat, branch_id = paths_to)

# get indices
idx <- dpt_flat@branch[, 1] \%in\% c(root, paths_to)

# get initial trajectory path
umap.path <- inferInitialTrajectory.v2(pt_vec[idx], so.query@reductions[["umap"]]@cell.embeddings[idx, ], w_width = 0.15)

   # get umap coordinates
df.umap.all <- data.frame(x = so.query@reductions[["umap"]]@cell.embeddings[ ,1],
                          y = so.query@reductions[["umap"]]@cell.embeddings[ ,2])
df.umap.all$col <- "grey"

df.umap.sub <- df.umap.all[idx, ]
df.umap.sub$pt <- pt_vec[idx]

# compute prinicpal curves
pc.fit <- princurve::principal_curve(as.matrix(df.umap.sub[,c("x", "y")]), start = as.matrix(umap.path[ ,c(1,2)]),
                                     thresh = 0.001, maxit = 10, stretch = 2, smoother = "smooth_spline",
                                     approx_points = 100, trace = FALSE, plot_iterations = FALSE)
traj.path <- pc.fit$s[pc.fit$ord, , drop = FALSE]
df.tp <- data.frame(traj.path)

}
