% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_functions.R
\name{inferSpecies}
\alias{inferSpecies}
\title{Infer species from list of Ensemble ids}
\usage{
inferSpecies(exp.mat, expected.species, rep.ens.method = "orig")
}
\arguments{
\item{exp.mat}{Expression matrix (genes x cell); row names are ensemble IDs, col names are cell IDs.}

\item{expected.species}{Character vector specifying which species are expected in expression matrix.
\itemize{
\item "Mm" - Mouse
\item "Hs" - Human
\item c("Mm", "Hs") - Mouse and Human.
}}

\item{rep.ens.method}{Method used to sample representative Ensemble ID for each cell (i.e., column).
\itemize{
\item "orig" - Original method; computation slower than alternative.
\item "alt" - Default. Alternative method;
}}
}
\value{
Character vector specifying species of each cell in expression  matrix.
}
\description{
For given gene expression matrix, assign species to each cell (column) based on Ensemble IDs.
}
