% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_functions.R
\name{initiateLog}
\alias{initiateLog}
\title{Initiate analysis log}
\usage{
initiateLog(module.name = "")
}
\arguments{
\item{module.name}{Character specifying module name}
}
\value{
data.frame
}
\description{
Create analysis log data.frame template
}
\examples{

# intiate data log
df.log <- initiateLog("log name")

# add new entry to log
df.log <- addLogEntry("Query File (.Rdata)", (input.file), df.log, "input.file")

}
\seealso{
\code{\link{addLogEntry}}
}
