% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_functions.R
\name{inititateLog}
\alias{inititateLog}
\title{Initiate analysis log}
\usage{
inititateLog(module.name = "")
}
\arguments{
\item{module.name}{Character specifying module name}
}
\value{
data.frame
}
\description{
Create analysis log data.frame template
}
