% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{isGeneAvailable}
\alias{isGeneAvailable}
\title{Check if gene is avaialble in Seurat Object}
\usage{
isGeneAvailable(so, query.gene, reference.genes)
}
\arguments{
\item{so}{Seurat Object}

\item{query.gene}{Character. Gene of interest.}

\item{reference.genes}{Named vector of genes; names are ENSEMBL, entries are SYMBOL.}
}
\value{
Logical
}
\description{
Return logical flag indicated whether query gene is present in seurat expression matrix.
}
