% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{lintersect}
\alias{lintersect}
\title{Returns intersection of all list entries.}
\usage{
lintersect(x)
}
\arguments{
\item{x}{list}
}
\description{
Returns intersection of all list entries.
}
\author{
Nicholas Mikolajewicz
}
