% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_functions.R
\name{loadCellRanger}
\alias{loadCellRanger}
\title{Load CellRanger preprocessed data}
\usage{
loadCellRanger(import_set, input_organisms, dir = "")
}
\arguments{
\item{import_set}{Character specifying folder containing matrix.mtx, gense.tsv and barcodes.tsv files provided by 10x.}

\item{input_organisms}{Character specifying species to include. This is necessary to parse ensemble ids into mouse and human. One of:
\itemize{
\item "Hs" - Human
\item "Mm" - Mouse
\item c("Hs", "Mm") - both species included
}}

\item{dir}{Character. folder containing import_set folder.}
}
\value{
list containing Seurat Object and named gene vector.
}
\description{
Input data is output from cell ranger (10x datasets)
}
