% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_functions.R
\name{loadMat}
\alias{loadMat}
\title{Load gene x cell count matrix into seurat object}
\usage{
loadMat(import_set, input_organisms, dir)
}
\arguments{
\item{import_set}{Character specifying matrix file name.}

\item{input_organisms}{All species included in input files. If multiple species are provided, error will thrown; only one expected. One of:
\itemize{
\item "Hs" - Human
\item "Mm" - Mouse
}}

\item{dir}{Character. folder containing import_set file}
}
\value{
list containing Seurat Object and named gene vector.
}
\description{
Load gene x cell count matrix into seurat object
}
