% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_functions.R
\name{loadTPM}
\alias{loadTPM}
\title{Load preprocessed TPM data (e.g., Neftel 2019 datasets)}
\usage{
loadTPM(import_set, subsample_factor, all_input_organisms, dir)
}
\arguments{
\item{import_set}{Character specifying TPM matrix file name.}

\item{subsample_factor}{Numeric [0,1]. Subsampling factor}

\item{all_input_organisms}{All species included in input files. If multiple species are provided, error will thrown; only one expected. One of:
\itemize{
\item "Hs" - Human
\item "Mm" - Mouse
}}

\item{dir}{Character. folder containing import_set file}
}
\value{
list containing Seurat Object and named gene vector.
}
\description{
Load preprocessed TPM data (e.g., Neftel 2019 datasets). While originally developed for TPM matrices, this function extends to any expression matrix.
}
