% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_functions.R
\name{m1.barcodeLabels}
\alias{m1.barcodeLabels}
\title{Assign barcodes labels to seurat metadata}
\usage{
m1.barcodeLabels(so, which.strata)
}
\arguments{
\item{so}{Seurat Object}

\item{which.strata}{Barcode labeling parameter. If NA, "subset_group" metadata field is set to "pooled".}
}
\value{
Seurat Object (with updated metadata)
}
\description{
Assign barcode labels (in Seurat object) to "subset_group" metadata field, as specified by which.strata parameter.
}
