% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_functions.R
\name{m1.filterSeurat}
\alias{m1.filterSeurat}
\title{Apply QC filters to Seurat Object}
\usage{
m1.filterSeurat(
  so,
  RNA.upperlimit = 9000,
  RNA.lowerlimit = 200,
  mt.upperlimit = 60,
  unmatch.low = 0,
  unmatch.high = 1
)
}
\arguments{
\item{so}{Seurat Object}

\item{RNA.upperlimit}{Upper limit threshold for genes/cell}

\item{RNA.lowerlimit}{Lower limit threshold for genes/cell}

\item{mt.upperlimit}{Numeric [0, 100]. Upper limit threshold for microchondiral content.}

\item{unmatch.low}{Numeric [0,1]. Lower limit threshold for unmatch rate. Default is 0. Ignored if data is not from sciRNA-seq3 pipeline.}

\item{unmatch.high}{Numeric [0,1]. Upper limit threshold for unmatch rate. Default is 1. Ignored if data is not from sciRNA-seq3 pipeline.}
}
\value{
List containing seurat object and filter summary statistics (as data.frame)
}
\description{
Filter scRNAseq data by gene/cell, unmatch rate and mitochondrial content.
}
