% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_functions.R
\name{m1.getMitoContent}
\alias{m1.getMitoContent}
\title{Calculate mitochondrial content}
\usage{
m1.getMitoContent(so, gNames, omit.na = T)
}
\arguments{
\item{so}{Seurat Object}

\item{gNames}{Named gene list; entries are Symbols, names are Ensemble.}

\item{omit.na}{Logical specifying whether to omit NA entries (present when unfiltered 10x dataset is used). Default is True.}
}
\value{
Seurat Object (with updated metadata)
}
\description{
Calculate cell-level mitochondrial content based on mapped transcripts and store in Seurat "percent.mt' metadata field.
}
