% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_functions.R
\name{m1.inferSpecies}
\alias{m1.inferSpecies}
\title{Infer species from list of Ensemble ids}
\usage{
m1.inferSpecies(exp.mat, expected.species, rep.ens.method = "alt")
}
\arguments{
\item{exp.mat}{Expression matrix (genes x cell); row names are ensemble IDs, col names are cell IDs.}

\item{expected.species}{Character vector specifying which species are expected in expression matrix.
\itemize{
\item "Mm" - Mouse
\item "Hs" - Human
\item c("Mm", "Hs") - Mouse and Human.
}}

\item{rep.ens.method}{Method used to sample representative Ensemble ID for each cell (i.e., column).
\itemize{
\item "orig" - Original method; computation slower than alternative.
\item "alt" - Default. Alternative method;
\item c("Mm", "Hs") - Mouse and Human.
}}
}
\value{
Seurat Object
}
\description{
For list of Ensemble ids, deconvolute species based on specified expectations.
}
