% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_functions.R
\name{m1.scNormScale}
\alias{m1.scNormScale}
\title{Normalize and Scale Data}
\usage{
m1.scNormScale(
  so,
  gNames,
  method = "SCT",
  var2regress = NULL,
  enable.parallelization = T,
  n.workers = 3,
  max.memory = (20480 * 1024^2)
)
}
\arguments{
\item{so}{Seurat Object}

\item{gNames}{Named vector of genes. Names are ensemble, entries are Symbols.}

\item{method}{Character specifying data normalization and scaling method. One of:
\itemize{
\item "NFS" - Seurat's NormalizeData, FindVariableFeatures, ScaleData workflow. Parameters are set to use LogNormalization method with a scale.factor of 1000. Variable features are selceted using 'mvp' method, and var2regress is regressed out during data scaling.
\item "SCT" - Default. SCTransform workflow; Uses regularized negative binomial regression to normalize UMI count data. All genes are returned (not only variable), and residual variace cutoff is set to 1.3; var2regress is included in model.
}}

\item{var2regress}{Character vector specifying which variables to regress out during data scaling.}

\item{enable.parallelization}{Logical specifying whether to enable parallelization. Default is T.}

\item{max.memory}{Max memory to use during parallel processing. Default is 20480 * 1024^2}

\item{n.works}{Number of works to used during parallel processing. Default is 3.}
}
\value{
Seurat Object
}
\description{
Applies one of two normalization/scaling approaches supported by Seurat.
}
