% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M23_functions.R
\name{m23.binarizeExpression}
\alias{m23.binarizeExpression}
\title{Stratify gene expression into high and low expression groups}
\usage{
m23.binarizeExpression(
  so,
  query.gene,
  quantile.threshold = 0.25,
  return.plt = F
)
}
\arguments{
\item{so}{Seurat Object}

\item{query.gene}{Character specifying gene of interest.}

\item{return.plt}{Logical indicating whether histogram with thresholds is returned. Otherwise Seurat object is returned.}

\item{quartile.threshold}{Quantile used to threshold expression data. E.g., 0.25 means that <25th percentile and >75th percentile are flagged as low and high expression groups, respecitvely.}
}
\value{
Seurat object (unless return.plt is T, in which case ggplot handle)
}
\description{
Creates new metadata field stratifying gene expression (for queried gene) into high and low expresion groups. If return.plt is true, histogram illustrating threshold is generated; Otherwise seurat object with new metadata entry is returned.
}
