% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_functions.R
\name{m9.importMarkers}
\alias{m9.importMarkers}
\title{M9. Import Gene List}
\usage{
m9.importMarkers(
  query.format,
  input.file = NULL,
  markers.of.interest = NULL,
  which.col = "first",
  which.species
)
}
\arguments{
\item{query.format}{Numerical code specifying gene list imput format. One of:
\itemize{
\item 1 - Character vector of genes. E.g., tnf.signalling = c("Tnfrsf1a", "Tnfrsf1b", "Tnfaip3")
\item 2 - csv file input.
}}

\item{input.file}{Character specifying module name. Required if query.format = 2.}

\item{markers.of.interest}{Character vector specifying list of gene(s). Required if query.format = 1.}

\item{which.col}{Character specifying which columsn to take gene lists from. Required if query.format = 2.
\itemize{
\item "first" - Include only first csv column.
\item "all" - Include all csv columns
}}

\item{which.species}{Character specifying species.}
}
\value{
List
}
\description{
Import gene lists for subsequent analysis in Module 9.
}
