% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{mapSubsetSeurat}
\alias{mapSubsetSeurat}
\title{Filter seurat object according to specified meta data entries}
\usage{
mapSubsetSeurat(so, old.field, new.field, mapping.list)
}
\arguments{
\item{so}{Seurat object.}

\item{old.field}{Existing seurat meta field that will be relabelled.}

\item{new.field}{New seurat meta field that will be created with relablled entries from old.field.}

\item{mapping.list}{Named list specifying how to relabel entries in old.field (mapping.list entries) to new.field (mapping.list names) in seurat meta data. Note that entries in mapping list do not have to match old.field entires exactly; entries are used as pattern arguemtn to grepl() function.}
}
\value{
Seurat object in which old field was mapped to new field, according to mapping specified in mapping.list. Any cells that are not mapped are omitted from seurat object.
}
\description{
Using a mapping.list, entries from an existing meta data field are relabeled and mapped to new meta data field, and cells which are not specified in this mapping are omitted from the seurat object.
}
\examples{

# define mapping lists
mapping.list.1 <- list(NSG = "NSG", BALBc = "BALB")
mapping.list.2 <- list(Early = "Early", Mid = "Mid", Late = "Late")

# Create new field called "Mouse" using relabled entries from "Condition" - relabelling based on mapping list
so.query.2 <- mapSubsetSeurat(so.query, old.field = "Condition", new.field = "Mouse", mapping.list = mapping.list.1)

# Create new field called "Time" using relabled entries from "Group" - relabelling based on mapping list
so.query.2 <- mapSubsetSeurat(so.query.2, old.field = "Group", new.field = "Time", mapping.list = mapping.list.2)

}
