% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{mergeSeuratList}
\alias{mergeSeuratList}
\title{Merge list of seurat objects}
\usage{
mergeSeuratList(so.list)
}
\arguments{
\item{so.list}{List of seurat objects}
}
\value{
Seurat Object
}
\description{
Merges list of seurat objects without any normalization of batch correction
}
