% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{miko_heatmap}
\alias{miko_heatmap}
\title{Function to draw ggplot heatmaps}
\usage{
miko_heatmap(
  mat,
  scale = "none",
  symmetric_scale = T,
  scale.lim = NA,
  color = colorRampPalette(rev(brewer.pal(n = 7, name = "RdBu")))(100),
  ...
)
}
\arguments{
\item{mat}{numeric matrix}

\item{scale}{character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. Corresponding values are "row", "column" and "none"}

\item{symmetric_scale}{Enforce symmetrical color scale. Default is TRUE.}

\item{scale.lim}{Apply ceiling and floor to all values in matrix. Default is NA.}

\item{color}{vector of colors used in heatmap}

\item{...}{additional parameters passed to pheatmap::pheatmap(...)}
}
\description{
Wrapper for pheatmap::pheatmap() and ggplotify::as.ggplot.
}
\seealso{
\code{\link{pheatmap}}
}
\author{
Nicholas Mikolajewicz
}
