% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{miko_integrate}
\alias{miko_integrate}
\title{scRNAseq integration wrapper}
\usage{
miko_integrate(
  object,
  split.by = "Barcode",
  min.cell = 50,
  k.anchor = 20,
  k.weight = 35,
  nfeatures = 3000,
  split.prenorm = F,
  assay = "SCT",
  variable.features.n = 3000
)
}
\arguments{
\item{object}{Seurat object}

\item{split.by}{Meta data feature to split and integrate data.}

\item{min.cell}{Minimum number of cells permitted per object prior to integration.}

\item{k.anchor}{How many neighbors (k) to use when picking anchors.}

\item{k.weight}{Number of neighbors to consider when weighting anchors.}

\item{nfeatures}{Number of features to return (passed to SelectIntegrationFeatures)}

\item{split.prenorm}{Split data before (TRUE) or after (FALSE) normalization.}

\item{assay}{Assay to use for normalization.}

\item{variable.features.n}{Use this many features as variable features after ranking by residual variance; default is 3000.}
}
\value{
Integrated seurat object
}
\description{
scRNAseq normalization and integration wrapper. Given seurat object input, data are split, normalized and integrated.
}
\examples{

}
\seealso{
\code{\link{IntegrateData}}
}
\author{
Nicholas Mikolajewicz
}
