% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{miko_message}
\alias{miko_message}
\title{Print message}
\usage{
miko_message(x, time = T, verbose = T)
}
\arguments{
\item{x}{Character string to print}

\item{time}{Logical to print time}

\item{verbose}{Logical to print message}
}
\description{
Print message
}
\author{
Nicholas Mikolajewicz
}
