% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{mikoScore}
\alias{mikoScore}
\title{Calculate Miko module scores for feature expression programs}
\usage{
mikoScore(
  object,
  geneset,
  nullscore,
  assay = DefaultAssay(object),
  nworkers = 1,
  nbin = 24,
  verbose = T
)
}
\arguments{
\item{object}{Seurat Object}

\item{geneset}{A list of vectors of features for expression programs; each entry should be a vector of feature names}

\item{nullscore}{`nullScore` output for provided `object`. Must run `nullScore` prior to running `mikoScore`.}

\item{assay}{Name of assay to use.}

\item{nworkers}{Number of workers for parallel implementation. Default is 1.}

\item{nbin}{Number of bins of aggregate expression levels for all analyzed features. See `AddModuleScore` for details.}

\item{verbose}{Print progress. Default is TRUE.}
}
\value{
list of results.
}
\description{
Feature expression programs are scored by first computing standardized module scores with `AddSModuleScore` and then scaling the scores by the null distributions calculated using `nullScore`. The resulting scores are robust to gene set size and can be further used to compute whether feature expression program is significantly up-regulated in single-cell population.
}
\seealso{
\code{\link{AddSModuleScore}} for standardized module scoring, \code{\link{nullScore}} for calculating null score distributions \code{\link{sigScore}} for derivation of p values for miko scores.
}
\author{
Nicholas Mikolajewicz
}
