% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_functions.R
\name{modulePackages}
\alias{modulePackages}
\title{Load Module-Specific Packages}
\usage{
modulePackages(module.number = NULL, load.default = T)
}
\arguments{
\item{module.number}{Numeric.}

\item{load.default}{Logical flag specifying wther to load default packages. If module.number is specified, load.default is ignored.}
}
\value{

}
\description{
Load packages required for each analysis module
}
