% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{multiSilhouette}
\alias{multiSilhouette}
\title{Evaluate silhouette indices of clustered single cell data across several cluster resolutions.}
\usage{
multiSilhouette(
  object,
  groups,
  assay_pattern = NULL,
  assay = NULL,
  verbose = T
)
}
\arguments{
\item{object}{Seurat object with multi-resolution clusters provided in meta data.}

\item{groups}{Vector specifying names of all cluster configurations found in meta data.}

\item{assay_pattern}{Cluster naming prefix.}

\item{assay}{Seurat assay used for clustering. If not specified, default assay is used.}

\item{verbose}{Print progress. Default is TRUE.}
}
\value{
Seurat object
}
\description{
Evaluate silhouette indices of clustered single cell data across several cluster resolutions. Consider running multiCluster(...) first.
}
\examples{
msil_list <- multiSilhouette(object = so.query, groups = cluster.name, assay_pattern = assay.pattern, verbose = T)
sil.plot <- msil_list$silhouette_plots
plt.silw.dep <- msil_list$resolution_plot
df.silw <- msil_list$silhouette_raw
df.silw.sum <- msil_list$silhouette_summary
rm(msil_list); invisible({gc()})
}
\seealso{
\code{\link{multiCluster}}
}
\author{
Nicholas Mikolajewicz
}
