% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{multiSpecificity}
\alias{multiSpecificity}
\title{Evaluate specificity of single-cell markers across several cluster resolutions.}
\usage{
multiSpecificity(
  object,
  cluster_names,
  features = NULL,
  deg_prefilter = T,
  cdi_bins = seq(0, 1, by = 0.01),
  min.pct = 0.1,
  n.workers = 1,
  return_dotplot = T,
  verbose = T
)
}
\arguments{
\item{object}{Seurat object with multi-resolution clusters provided in meta data.}

\item{cluster_names}{Vector specifying names of all cluster configurations found in meta data.}

\item{features}{If specified, marker specificity analysis is limited to specified features. Otherwise all features are used (more computationally intensive).}

\item{deg_prefilter}{If TRUE, wilcoxon analysis is performed first to subset DEG features for downstream analysis. Results in faster performance. Default is TRUE.}

\item{cdi_bins}{Vector specifying binning for CDI-based specificity curve. Must range [0,1]. Default is seq(0, 1, by = 0.01).}

\item{min.pct}{Minimal expression of features that are considered in specificity analysis. Represents fraction of expression cells and must range [0,1]. Higher values result in faster performance. Default is 0.1.}

\item{n.workers}{Number of workers used for parallel implementation. Default is 1.}

\item{return_dotplot}{If TRUE, dot plots visualizing expression of top specific markers are returned. Default is T.}

\item{verbose}{Print progress. Default is TRUE.}
}
\value{
Seurat object
}
\description{
Evaluate specificity of single-cell markers across several cluster resolutions using co-dependency index-based specificity measure. Consider running multiCluster(...) first.
}
\examples{
# clustering data
ms.list <- multiSpecificity(object = so.query, cluster_names = cluster.name, features = NULL, deg_prefilter = T,
cdi_bins = seq(0, 1, by = 0.01), min.pct = 0.1,
n.workers = 4, return_dotplot = T,  verbose = T)
df.auc.spec <- ms.list$specificity_summary
qm.res.sum.sum.all <- ms.list$specificity_raw
plt.clust.spec <- ms.list$auc_plot
plt.auc.spec <- ms.list$resolution_plot
plt.auc.dot <- ms.list$dot_plot
}
\seealso{
\code{\link{multiCluster}}
}
\author{
Nicholas Mikolajewicz
}
