% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{namedList2longDF}
\alias{namedList2longDF}
\title{Convert named list to long data.frame}
\usage{
namedList2longDF(my.list, name.header = NULL, value.header = NULL)
}
\arguments{
\item{my.list}{named list}

\item{name.header}{character specifying header name that will be assigned to name column. If unspecified, defaults to "name".}

\item{value.header}{character specifying header name that will be assigned to value column. If unspecified, defaults to "value".}
}
\value{
long data.frame
}
\description{
Convert named list to long data.frame. Resulting dataframe will have two columns, the first corresponding to the names within the list, and the second to the corersponding list entries.
}
\examples{

# specify named list
my.list <- list(group.1 = c("a", "b", "c"), group.2 = c("d", "e", "f"))

# convert named list to long data.frame
my.df <- namedList2longDF(my.list)

}
\seealso{
\code{\link{namedList2wideDF}}
}
\author{
Nicholas Mikolajewicz
}
