% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{namedList2wideDF}
\alias{namedList2wideDF}
\title{Convert named list to wide data.frame}
\usage{
namedList2wideDF(my.list)
}
\arguments{
\item{my.list}{named list}
}
\value{
wide data.frame
}
\description{
Convert named list to wide data.frame. Resulting dataframe will have the same number of columns as there are names in the list.
}
\examples{

# specify named list
my.list <- list(group.1 = c("a", "b", "c"), group.2 = c("d", "e", "f"))

# convert named list to wide data.frame
my.df <- namedList2wideDF(my.list)

}
