% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{neighborPurity}
\alias{neighborPurity}
\title{Compute purity of each cell's neighborhood, as defined by KNN graph.}
\usage{
neighborPurity(object, graph, cluster.field = "seurat_clusters", verbose = T)
}
\arguments{
\item{object}{Seurat object}

\item{graph}{name of KNN graph to use for analysis. Must be present within provided seurat object. If absent, run FindNeighbors().}

\item{cluster.field}{name of cluster metadata field to use for cluster membership data. Default is 'seurat_clusters'.}

\item{verbose}{Print progress. Default is TRUE.}
}
\value{
Seurat object with purity score stored in metadata
}
\description{
Compute purity of each cell's neighborhood, as defined by KNN graph. For each neighborhood, the proportion of cells belonging to the most represented cluster is computed as the purity score. The Purity score is stored in the metadata of the returned seurat object.
}
\examples{

object <- FindNeighbors(object)
object <- FindClusters(object, resolution = 1)
object <- neighborPurity(object, "RNA_nn")

}
\author{
Nicholas Mikolajewicz
}
