% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{netHG}
\alias{netHG}
\title{Construct functional enrichment network}
\usage{
netHG(
  hg.res,
  gene.list,
  species,
  fdr.filter = 0.1,
  edge.threshold = 0.25,
  cex_line = 0.2,
  pie = "Count",
  gene.universe = NULL,
  verbose = T
)
}
\arguments{
\item{hg.res}{Output from runHG(...)}

\item{gene.list}{Named list of genesets used in runHG(...).}

\item{species}{Species. One of "Mm" (mouse) or "Hs" (human)}

\item{fdr.filter}{Significance threshold. Default is 0.1.}

\item{edge.threshold}{Edge threshold for visualization. Default is 0.25.}

\item{cex_line}{Edge size. Default is 0.2.}

\item{gene.universe}{Gene universe (symbol format). If not specified, uses all genes provided in gene.list.}

\item{verbose}{Print progress. Default is TRUE.}
}
\value{
enrichment results
}
\description{
Construct functional enrichment network
}
\seealso{
\code{\link{runHG}}
}
\author{
Nicholas Mikolajewicz
}
