% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{networkProperties.Plot}
\alias{networkProperties.Plot}
\title{Plot network properties for different soft thresholds}
\usage{
networkProperties.Plot(sft, r2.threshold = 0.85)
}
\arguments{
\item{sft}{list of soft threshold picks (output from scMiko::getSoftThreshold)}

\item{r2.threshold}{Numeric specifying threshold at which to draw red horizontal curve. Default is 0.85.}
}
\value{
plot
}
\description{
Plot network properties for different soft thresholds
}
\examples{

sft <- getSoftThreshold(s.mat, networkType = "unsigned")
networkProperties.Plot(sft)


}
