% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{nullScore}
\alias{nullScore}
\title{Null distribution for standardized module scores.}
\usage{
nullScore(
  object,
  assay = DefaultAssay(object),
  n.replicate = 25,
  min.gs.size = 2,
  max.gs.size = 100,
  step.size = 2,
  nworkers = 1,
  subsample.n = NULL,
  nbin = 24,
  verbose = T
)
}
\arguments{
\item{object}{Seurat Object}

\item{assay}{Name of assay to use}

\item{n.replicate}{Number of replicates}

\item{min.gs.size}{minimum gene set size}

\item{max.gs.size}{maximum gene set size}

\item{step.size}{step size; increment of sequence between `min.gs.size` and `max.gs.size`.}

\item{subsample.n}{Numeric [1,ncol(object)]. Number of cells to subsample. If specified, overides subsample.factor.}

\item{nbin}{Number of bins of aggregate expression levels for all analyzed features. See `AddModuleScore` for details.}

\item{verbose}{Print progress. Default is TRUE.}
}
\value{
list of results that are used as input into `mikoScore`.
}
\description{
Downsample number of cells in Seurat object by specified factor
}
\seealso{
\code{\link{AddSModuleScore}} for standardized module scoring, \code{\link{mikoScore}} for miko scoring, \code{\link{sigScore}} for derivation of p values for miko scores.
}
\author{
Nicholas Mikolajewicz
}
