% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{optimalBinSize}
\alias{optimalBinSize}
\title{Identify optimal bin size for AddModuleScore() function}
\usage{
optimalBinSize(object, pool = NULL, nbin = 24, seed = 1023, verbose = T)
}
\arguments{
\item{object}{Seurat object}

\item{pool}{List of features to check expression levels against. Defaults to rownames(x = object).}

\item{seed}{See a random seed. If NULL, seed is not set.}

\item{nbinNumber}{of bins of aggregate expression levels for all analyzed features. Initial value to begin with looking for optimal bin.}
}
\description{
Identify optimal bin size for AddModuleScore() function
}
\examples{
optimal.nbin <-  optimalBinSize(object = so.query)
}
\author{
Nicholas Mikolajewicz
}
