% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{orderedFactor}
\alias{orderedFactor}
\title{Sort factor levels in numerical order}
\usage{
orderedFactor(f)
}
\arguments{
\item{factor}{factor with numerical entries.}
}
\value{
ordered factor
}
\description{
Sort factor levels in numerical order
}
\author{
Nicholas Mikolajewicz
}
