% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{p2z}
\alias{p2z}
\title{Convert p value to z score}
\usage{
p2z(p)
}
\arguments{
\item{p}{p value}
}
\value{
z score (absolute value)
}
\description{
Convert p value to z score
}
\seealso{
\code{\link{qnorm}}
}
