% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{parCor}
\alias{parCor}
\title{Parallelized correlation}
\usage{
parCor(mat, method = "spearman", do.par = T, n.workers = 4)
}
\arguments{
\item{method}{Correlation method, a character. Options: 'pearson' or 'spearman'.}

\item{do.par}{Logical indicating whether to use parallel implementation. Default is T.}

\item{n.workers}{number of workers to use for parallel computation.}

\item{max}{matrix}
}
\description{
Parallelized correlation
}
\author{
Nicholas Mikolajewicz
}
