% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pathActivity.Tempora}
\alias{pathActivity.Tempora}
\title{Calcualte pathway enrichment scores in Tempora Framework}
\usage{
pathActivity.Tempora(
  object,
  gmt_path,
  method = "gsva",
  min.sz = 5,
  max.sz = 200,
  parallel.sz = 1,
  verbose = F,
  do.plot = F
)
}
\arguments{
\item{object}{Tempora Object}

\item{gmt_path}{Local path to database of pathways or genesets organized as a .gmt file. Genesets files in GMT format can be downloaded at http://baderlab.org/GeneSets.}

\item{method}{Method used to estimate pathway enrichment profile per cluster. Can be "gsva", "ssgsea", "zscore" or "plage", default to "gsva". See ?gsva for more information.}

\item{min.sz}{Minimum size of the genesets used in enrichment estimation, set to 5 genes by default.}

\item{max.sz}{Maximum size of the genesets used in enrichment estimation, set to 200 genes by default.}

\item{parallel.sz}{Type of cluster architecture when using snow. If 1, no parallelization will be used. If 0, all available cores will be used.}

\item{verbose}{Flag for reporting progress}

\item{do.plot}{Flag for ploting scree plot}
}
\value{
Tempora object with internally calculated pathway scores
}
\description{
Calcualte pathway enrichment scores in Tempora Framework. Uses GSVA backend.
}
\examples{

}
\author{
Gary Bader et. al
}
