% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prepExpression}
\alias{prepExpression}
\title{Normalize and scale data within appropriate assay}
\usage{
prepExpression(so)
}
\arguments{
\item{so}{Seurat Object}
}
\value{
Seurat Object
}
\description{
Set correct assay and prep expression data for downstream analysis. This handles both intergrated and non-integrated datasets. If input is integrated dataset, set assay to RNA and normalized and resscale data. If input is non-integrated dataset, us scTransformed data for expression analysis.
}
