% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prepGeneList}
\alias{prepGeneList}
\title{Prepare gene to ensemble conversion vector}
\usage{
prepGeneList(so, global.enviroment, species = "Hs")
}
\arguments{
\item{so}{Seurat Object}

\item{global.enviroment}{global.enviroment <- objects()}
}
\value{
Named vector of available genes
}
\description{
Prepare gene to ensemble conversion vector
}
\examples{

gNames.list <- prepGeneList(so, objects())

}
\author{
Nicholas Mikolajewicz
}
