% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prepSeurat}
\alias{prepSeurat}
\title{prep Seurat}
\usage{
prepSeurat(object)
}
\arguments{
\item{object}{Seurat objects}
}
\value{
Seurat object
}
\description{
Preprocess using fixBarcodeLabel(), UpdateSeuratObject() and updateDimNames() Functions.
}
\author{
Nicholas Mikolajewicz
}
