% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{projectReduction}
\alias{projectReduction}
\title{Project dimensionally-reduced features onto UMAP}
\usage{
projectReduction(
  object,
  reduction = "pca",
  n.components = NA,
  show.n.features = 50,
  pca.min.var.exp = 0.05,
  pca.cum.pca.thresh = 0.8,
  umap.reduction = "umap",
  rel.width = c(2.5, 1),
  ...
)
}
\arguments{
\item{object}{Seurat Object}

\item{reduction}{Character specifying reduction key. Default: "pca"}

\item{n.components}{Numeric. Number of top components to project.}

\item{show.n.features}{Numeric. Show top n features alongside projection. Default is 50.}

\item{pca.min.var.exp}{Threshold for minimum variance explained by principal component.}

\item{pca.cum.pca.thresh}{Threshold for maximum cumulative variance explained by principal component.}

\item{umap.reduction}{Character specifying reduction key for umap coordinates.}

\item{rel.width}{Relative widths of umap and top feature plots. Default: c(2.5,1)}

\item{...}{additional parameters passed to geom_point(). Generally used to specify point size.}
}
\value{
list of ggplot handles
}
\description{
Project dimensionally-reduced features (e.g., PC, IC) onto UMAP
}
\examples{


}
\author{
Nicholas Mikolajewicz
}
