% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{propVarPCA}
\alias{propVarPCA}
\title{Variance explained by each principal component.}
\usage{
propVarPCA(so, reduction.name = "pca")
}
\arguments{
\item{so}{Seurat Object}

\item{reduction.name}{Name of reduction to use. Default is 'pca'.}
}
\value{
data.frame summarize proportion of variance explained by each principal component
}
\description{
For given Seurat Object, retrieve principal components and compute proportion of explained variance.
}
\author{
Nicholas Mikolajewicz
}
