% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{pruneSSN}
\alias{pruneSSN}
\title{Identify and (optionally) prune gene program features in scale-free shared nearest neighbor network (SSN)}
\usage{
pruneSSN(object, graph = "RNA_snn_power", prune.threshold = 0.1, return.df = F)
}
\arguments{
\item{object}{Seurat object objected from SSN analysis.}

\item{graph}{name of shared nearest neighbor (SNN) graph used for SSN analysis.}

\item{prune.threshold}{threshold used for pruning features from SSN graph (features with low connectivity are preferentially pruned). Default is 0.1.}

\item{return.df}{whether to return data.frame instead of named list.}
}
\value{
Returns named list of SSN gene program features.
}
\description{
Identify and (optionally) prune gene program features in scale-free shared nearest neighbor network (SSN)
}
\examples{

}
\seealso{
\code{\link{runSSN}} for SSN analysis
}
\author{
Nicholas Mikolajewicz
}
