% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pseudoReplicates}
\alias{pseudoReplicates}
\title{Create pseudo-replicates, stratified by grouping variable.}
\usage{
pseudoReplicates(object, split.by, n = 2)
}
\arguments{
\item{object}{Seurat object}

\item{split.by}{name of metadata column to create pseudo-replicates for.}

\item{n}{numerical, number of replicates to create.}
}
\value{
Returns a Seurat object where latest pseudo replicates results will be stored in object metadata under 'pseudo_replicates'. Note that 'pseudo_replicates' will be overwritten every time pseudoReplicates is run
}
\description{
Create pseudo-replicates, stratified by grouping variable.
}
\author{
Nicholas Mikolajewicz
}
