% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{pseudotime.UMAP}
\alias{pseudotime.UMAP}
\title{UMAP projection of pseudotimes}
\usage{
pseudotime.UMAP(
  x,
  y,
  pseudotime,
  pt.size = 1,
  pt.alpha = 1,
  x.lab = "UMAP 1",
  y.lab = "UMAP 2"
)
}
\arguments{
\item{x}{x coordinates (e.g., UMAP 1)}

\item{y}{y coordinates (e.g., UMAP 2)}

\item{pseudotime}{pseudotimes. Numerical vector, same length as x and y.}

\item{pt.size}{Point size. Numeric [0,inf]}

\item{pt.alpha}{Point alpha. Numeric [0,1]}

\item{x.lab}{x axis label. Default is 'UMAP 1'.}

\item{y.lab}{y axis label. Default is 'UMAP 2'.}
}
\value{
ggplot2 handle
}
\description{
Visualize pseudotimes using UMAP coordinates.
}
\examples{


}
