% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pseudotimeRF}
\alias{pseudotimeRF}
\title{Identify pseudotime-dependent genes using Random Forest (RF) Model}
\usage{
pseudotimeRF(
  so,
  hvg,
  pseudotimes,
  lineage.name,
  slot = "data",
  assay = DefaultAssay(so),
  mtry = length(hvg)/10,
  trees = 1000,
  min_n = 15,
  mode = "regression",
  importance = "impurity",
  num.threads = 3
)
}
\arguments{
\item{so}{Seurat Object}

\item{hvg}{Genes used to fit model (character vector; must be available in rows of seurat object). It is suggested to keep number of genes low (~200) for optimal performance.}

\item{pseudotimes}{Numeric vector of pseudotimes. Length must be equal to number of cells in seurat object (ncol(so)).}

\item{lineage.name}{Name of pseudotime lineage; used to label results.}

\item{slot}{A character specifying which slot to pull data from; default is 'Data'}

\item{assay}{A character specifying which assay to use (e.g., 'RNA' or 'SCT'). If unspecified, set to DefaultAssay(so)}

\item{mtry}{An integer for the number of predictors that will be randomly sampled at each split when creating the tree models.}

\item{trees}{An integer for the number of trees contained in the ensemble.}

\item{min_n}{An integer for the minimum number of data points in a node that are required for the node to be split further.}

\item{mode}{Specfiy type of RF to fit: 'regression' or 'classification'. Regression is default and it is not recommended to change this argument.}

\item{importance}{Type of importance. Default is 'impurity'.}

\item{num.threads}{An integer for the number of threads to use when fitting RF model}
}
\value{
List of results
}
\description{
Identify pseudotime-dependent genes using Random Forest (RF) Model. Given Seurat object and highly variable genes (hvg), expression data are split into training and test set, and random forest model is trained to identify genes that vary with pseudotime. Model performance is evaluated on test set. Random forest model is fit using R 'parsnip' package.
}
\seealso{
\code{\link{rand_forest}}
}
