% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{QC.scatterPlot}
\alias{QC.scatterPlot}
\title{QC scatter plots}
\usage{
QC.scatterPlot(so)
}
\arguments{
\item{so}{Seurat Object}
}
\value{
ggplot handle
}
\description{
Pairwise relationships between 1) UMI count/cell and mitchondrial content, 2) gene count/cell and mitochondrial content, and 3) UMI count/cell and gene count/cell.
}
