% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{QC.violinPlot}
\alias{QC.violinPlot}
\title{QC violin plots}
\usage{
QC.violinPlot(so, plt.log.flag = T)
}
\arguments{
\item{so}{Seurat Object}

\item{plt.log.flag}{Logical specifying whether data are plotted on log scale. Default is True.}
}
\value{
list of ggplot handles
}
\description{
Number of genes/cell, UMI/cell and mitochondiral content/cell are visualized with violin plots. Two ggplot handles are generated. First contains QC metrics pooled across all cells, while second stratifies dataset by barcode labels.
}
