% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{qNorm}
\alias{qNorm}
\title{Quantile Normalization of 2 Vectors}
\usage{
qNorm(x, y, genes = NULL, flag.top.n = 15)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}

\item{genes}{Character vector of gene names. Used to label entries in data.frame output.}

\item{flag.top.n}{Numeric indicating top n genes to flag in data.frame output. Default is 15.}
}
\value{
data.frame of quantile normalized values.
}
\description{
Performs quantile normalization of 2 vectors (Hicks 2014).
}
\examples{

# quantile normalization
x.s1 <- getConnectivity(s.1, gene.names = colnames(a.1))$wi
y.s2 <- getConnectivity(s.2, gene.names = colnames(a.2))$wi
df.xy.s <- qNorm(x.s1, y.s2, genes = colnames(a.1))

}
