% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{recaleValues}
\alias{recaleValues}
\title{Rescale values to specified range.}
\usage{
recaleValues(values, new.min = 0, new.max = 1)
}
\arguments{
\item{values}{Numeric vector of values to rescale.}

\item{new.min}{Numeric specfiying new minimum. If unspecified, default is 0.}

\item{new.max}{Numeric specfiying new maximum. If unspecified, default is 1.}
}
\value{
Numeric vector of rescaled values
}
\description{
Rescales vector of values to span between specified range [new.min, new.max]
}
\examples{

# rescale values
values <- recaleValues(values)

}
