% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_functions.R
\name{recodeBarcode}
\alias{recodeBarcode}
\title{Recode (i.e., relabel) metadata in Seurat object}
\usage{
recodeBarcode(
  object,
  barcode.recode,
  old.field.name = "Barcode",
  new.field.name = "Barcode"
)
}
\arguments{
\item{object}{Seurat Object}

\item{barcode.recode}{Named list where names are new labels, entries are old labels (can be vector of characters). E.g., barcode.recode <- list(in.vitro = c("invitro"), in.vivo = NA)}

\item{old.field.name}{name of field in which meta dat for recoding is stored. Default is "Barcode".}

\item{new.field.name}{name of new field in meta data where recoded data will be stored. Default is "Barcode".}
}
\value{
Seurat Object with relabeled metadata
}
\description{
Recode (i.e., relabel) metadata in Seurat object. If old label is specified as NA, treated as wildcard that captures all remaining barcodes that have not been relabeled.
}
\author{
Nicholas Mikolajewicz
}
