% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rmDuplicateGenes}
\alias{rmDuplicateGenes}
\title{Remove duplicate genes from Seurat Object}
\usage{
rmDuplicateGenes(object, retain.graphs = T)
}
\arguments{
\item{object}{Seurat Object}

\item{retain.graphs}{Logical specifying whether to retain graphs in subset Seurat object because Seurat::subset function inherently removed the graphs. Default is TRUE.}
}
\value{
Seurat object
}
\description{
Remove duplicate genes from Seurat Object.
}
\examples{

so.query <- rmDuplicateGenes(so.query, retain.graphs = T)

}
\author{
Nicholas Mikolajewicz
}
