% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rmvCSVprefix}
\alias{rmvCSVprefix}
\title{Remove "ï.." prefix that is appended to csv header}
\usage{
rmvCSVprefix(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character vector
}
\description{
Remove "ï.." prefix that is appended to csv header
}
