% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{runGSEA}
\alias{runGSEA}
\title{Run gene-set enrichment analysis (GSEA)}
\usage{
runGSEA(
  gene,
  value,
  species,
  db = "GO",
  my.entrez = NULL,
  my.pathway = NULL,
  min.size = 3,
  max.size = 300,
  do.plot = T,
  plot.top.n = 10,
  path.name.wrap.width = 40
)
}
\arguments{
\item{gene}{Character vector of gene names}

\item{value}{Numeric vector of values used to rank genes (e.g., logFC, expression, etc.). Must be same length as gene argument.}

\item{species}{Species. One of "Mm" (mouse) or "Hs" (human)}

\item{my.entrez}{Data.frame with SYMBOL and ENTREZID columns. Used to relate gene symbols to Entrez IDs. Retrieved internally unless explicitly provided.}

\item{my.pathway}{Named list of pathways, with each entry containing vector of Entrez IDs. Retrieved internally unless explicitly provided.}

\item{min.size}{Minimum gene set size. Default is 3.}

\item{max.size}{Minimum gene set size. Default is 300.}

\item{do.plot}{Logical to return dotplot visualizing top GSEA ranked pathways. Default is T.}

\item{plot.top.n}{Numeric specifying how many top pathways to visualize. Default is 10.}

\item{path.name.wrap.width}{Numeric specifying width of pathway names to wrap around. Argument passed to stringr::str_wrap(..., width = path.name.wrap.width)}
}
\description{
Run gene-set enrichment analysis (GSEA)
}
\author{
Nicholas Mikolajewicz
}
