% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{runHG}
\alias{runHG}
\title{Run hypergeometric gene enrichment analysis.}
\usage{
runHG(gene.list, gene.universe, species, pathway.db = "Bader", n.workers = 16)
}
\arguments{
\item{gene.list}{Named list of genesets to run enrichment on (symbol format).}

\item{gene.universe}{Gene universe (symbol format)}

\item{species}{Species. One of "Mm" (mouse) or "Hs" (human)}

\item{pathway.db}{Name of databse to get annotation genelists from. Default is "Bader". See scMiko::getAnnotationPathways() for options.}

\item{n.workers}{Number of workers for parallelization. Default is 16.}
}
\description{
Run hypergeometric gene enrichment analysis.
}
\author{
Nicholas Mikolajewicz
}
