% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{runICA}
\alias{runICA}
\title{Run Independent Component Analysis on gene expression}
\usage{
runICA(object, features = NULL, max_cells = 20000, verbose = T)
}
\arguments{
\item{object}{ICA feature loadings}

\item{features}{Features to compute ICA on. If not specified, all features present in `object` are used (not recommended, ICA is computationally expensive).}

\item{max_cells}{Max number of cells to run ICA on. If number of cells in `object` exceeded `max_cells`, object is subsampled prior to running ICA. Default is 20000.}

\item{verbose}{Print progress. Default is T.}
}
\value{
seurat object with significant genes stored in "misc" slot of ICA reduction slot.
}
\description{
Run fastica algorithm from the ica package for ICA dimensionality reduction. Wrapper for Seurat's RunICA function, with additional significant gene identification step.
}
\examples{

}
\seealso{
\code{\link{RunICA}} for Seurat's independent component analysis, \code{\link{getICAGenes}} for significant ICA gene identification.
}
\author{
Nicholas Mikolajewicz
}
