% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{runMS}
\alias{runMS}
\title{Run Modular Scoring}
\usage{
runMS(
  object,
  genelist,
  assay = DefaultAssay(object),
  score.key = "MS",
  size = autoPointSize(ncol(object)),
  raster = F,
  rescale = F,
  verbose = T,
  winsorize.quantiles = c(0, 1),
  search = F,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{genelist}{Named list of genesets.}

\item{assay}{Assay used for expression matrix.}

\item{score.key}{Expression program prefix. default is "MS".}

\item{size}{UMAP point size.}

\item{raster}{Convert points to raster format, default is FALSE.}

\item{rescale}{rescale values from 0 to 1. Default is FALSE.}

\item{verbose}{Print progress. Default is TRUE.}

\item{winsorize.quantiles}{Rescale values to lie between lower and upper bound quanitle. Default  = c(0,1).}

\item{search}{Search for symbol synonyms for features in features that don't match features in object? Searches the HGNC's gene names database; see UpdateSymbolList for more details. Default is FALSE.}

\item{...}{additional parameters passed to geom_point(...)}
}
\value{
list of results along with ggplot handles visualizing class predictions overlaid on UMAP.
}
\description{
Wrapper for Seurat::AddModuleScore(). Calculates module scores for feature expression programs in single cells.
}
\examples{

# get genesets
verhaak.df <- geneSets[["Verhaak_CancerCell_2010"]]
verhaak.list <- wideDF2namedList(verhaak.df)

gsc.df <- geneSets[["Richards_NatureCancer_2021_sc"]]
gsc.list <- wideDF2namedList(gsc.df)

neftel.df <- geneSets[["GBM_Hs_Neftel2019"]]
neftel.list <- wideDF2namedList(neftel.df)

verhaak.list <- lapply(verhaak.list, toupper)
gsc.list <- lapply(gsc.list, toupper)
neftel.list <- lapply(neftel.list, toupper)

# classify cells based on provided genesets
v.auc <- runMS(object = so.query, genelist = verhaak.list)
v.auc$plot.max.score

g.auc <- runMS(object = so.query, genelist = gsc.list)
g.auc$plot.max.score

n.auc <- runMS(object = so.query, genelist = neftel.list)
n.auc$plot.max.score

}
\seealso{
\code{\link{AddModuleScore}}
}
\author{
Nicholas Mikolajewicz
}
