% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{runNMF}
\alias{runNMF}
\title{Perform non-negative matrix factorization (NMF)}
\usage{
runNMF(
  object,
  assay = DefaultAssay(object),
  k = 6,
  raster = F,
  n.threads = 2,
  max.iter = 50,
  gene.cutoff = 0.5,
  gene.n = 50,
  sample.name = NULL,
  show.top.n = 10,
  min.pct = NA,
  verbose = T
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{assay. Default "SCT".}

\item{k}{Number of NMF modules. Default is 6.}

\item{raster}{rasterize output plot. Default is F.}

\item{n.threads}{Number of threads to use when running NMF.}

\item{max.iter}{FDR threshold. Default = 0.01.}

\item{gene.cutoff}{gene cutoff threshold [0,1]. Default = 0.5. Ignored if gene.n is specified.}

\item{gene.n}{number of genes to return per module.}

\item{sample.name}{sample name. Default is NULL.}

\item{show.top.n}{number of enrichment terms to show in summary plots.}

\item{min.pct}{gene expression threshold [0, 1]. Default is NA.}

\item{verbose}{Print progress. Default is TRUE.}
}
\value{
list
}
\description{
Perform non-negative matrix factorization (NMF)
}
\examples{

}
\seealso{
\code{\link{nnmf}}
}
\author{
Nicholas Mikolajewicz
}
