% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{runRPCA}
\alias{runRPCA}
\title{Run Robust Prinicipal Component Analysis}
\usage{
runRPCA(
  object,
  assay = NULL,
  features = NULL,
  npcs = 50,
  maxpcs = 50,
  reduction.key = "RPC_",
  reduction.name = "rpca",
  seed.use = 42,
  verbose = T,
  method = c("hubert", "robpca", "fasthcs", "pcal"),
  maxdir = 100,
  signflip = T,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Name of Assay rPCA is being run on}

\item{features}{Features to compute PCA on. If features=NULL, PCA will be run using scaled features for the Assay. Note that the features must be present in the scaled data. Any requested features that are not scaled or have 0 variance will be dropped, and the PCA will be run using the remaining features.}

\item{npcs}{Total Number of PCs to compute and store (50 by default)}

\item{maxpcs}{Max Number of PCs to compute and store (50 by default)}

\item{reduction.name}{dimensional reduction name, rpca by default}

\item{seed.use}{Set a random seed. By default, sets the seed to 42. Setting NULL will not set a seed.}

\item{verbose}{Print progress. Default is TRUE.}

\item{method}{Robust PCA method. default is "hubert".}

\item{maxdir}{maximal number of random directions to use for computing the outlyingness of the data points. Default is maxdir=100.}

\item{signflip}{a logical value indicating wheather to try to solve the sign indeterminancy of the loadings - ad hoc approach setting the maximum element in a singular vector to be positive. Default is signflip = TRUE}

\item{...}{additional parameters passed to rPCA methods.}

\item{reductions.key}{dimensional reduction key, specifies the string before the number for the dimension names. RPC by default}
}
\value{
Seurat object
}
\description{
Run a robust PCA (rPCA) dimensionality reduction on single-cell seurat object.
}
\seealso{
\code{\link{PcaHubert}}
}
\author{
Nicholas Mikolajewicz
}
