% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_functions.R
\name{runScanorama}
\alias{runScanorama}
\title{Integrate scRNA-seq data using Scanorama}
\usage{
runScanorama(
  object,
  batch,
  do.umap = F,
  nfeatures = 2000,
  vars2regress = "percent.mt",
  dim = 50L,
  verbose = T
)
}
\arguments{
\item{object}{Seurat object or list of seurat objects.}

\item{batch}{name of meta data field specifying batch groupings.}

\item{do.umap}{compute UMAP. Default is T.}

\item{nfeatures}{number of features to use for integration. Default is 2000.}

\item{vars2regress}{variable to regress out. Ignored if SCT assay is available in provided object.}

\item{dim}{number of dimensions to use for integration. Default is 50L.}

\item{verbose}{print progress. Default is T.}
}
\value{
Seurat object.
}
\description{
This is a wrapper for Scanorama (https://github.com/brianhie/scanorama). Scanorama enables batch-correction and integration of heterogeneous scRNA-seq datasets.
}
\examples{
 object <- runScanorama(object, batch = "sample")

}
\author{
Nicholas Mikolajewicz
}
