% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn_functions.R
\name{runWNN}
\alias{runWNN}
\title{Run WNN Multi-Modal Integration.}
\usage{
runWNN(
  object,
  do.scale = F,
  do.center = F,
  pca.thres = 0.9,
  cres = 1,
  min.pct = 0.25,
  split.var = "seurat_clusters"
)
}
\arguments{
\item{object}{Seurat object or list of expression matrices. If seurat object, expression matrices are extracted. If list, assumes that expression matrix entries have column-wise genes and row-wise cells.}

\item{do.scale}{Logical to scale expression. Default is F.}

\item{do.center}{Logical to center expression. Default is F.}

\item{pca.thres}{Variance explained threshold for PC component inclusion. Default is 0.9.}

\item{cres}{Cluster resolution for integrated network. Default is 1.}

\item{min.pct}{Minimum expression fraction for inclusion in network integration. Default is 0.25.}

\item{split.var}{Grouping variable for expression fraction filter. Default is 'seurat_clusters'}
}
\value{
list of integrated results
}
\description{
Run WNN Multi-Modal Integration. Modified wrapper for seurat WNN workflow.
}
\author{
Nicholas Mikolajewicz
}
