% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{savePDF}
\alias{savePDF}
\title{Save figure as pdf}
\usage{
savePDF(file.name, plot.handle, fig.width = 5, fig.height = 5, save.flag = T)
}
\arguments{
\item{file.name}{path/filename of output pdf. A character.}

\item{plot.handle}{plot handle that will be saved as pdf. E.g., GGplot handle.}

\item{fig.width}{Numeric for plot width. Default is 5.}

\item{fig.height}{Numeric for plot height Default is 5.}

\item{save.flag}{Logical to save figure. Default is TRUE.}
}
\description{
Save figure as pdf
}
\examples{

# save module figure
savePDF(file.name = "M01_QC_violin.pdf", plot.handle = plt.QC_violin, fig.width = 5, fig.height = 5, save.flag = save.pdf)

}
\author{
Nicholas Mikolajewicz
}
