% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{savePlot}
\alias{savePlot}
\title{Save ggplot as image file}
\usage{
savePlot(plt.handle, save.as, format = "png", ...)
}
\arguments{
\item{plt.handle}{ggplot handle}

\item{save.as}{save file name. A character.}

\item{format}{Format of saved image
\itemize{
\item "png" - Default
\item "pdf"
\item "jpeg"
\item "ps"
\item "bmp"
\item "wmf"
}}

\item{...}{pass additional arguments to graphics device functions.}
}
\value{

}
\description{
Export ggplot to image file
}
