% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{scale_color_miko}
\alias{scale_color_miko}
\title{Gradient color scale}
\usage{
scale_color_miko(
  low = scales::muted("blue"),
  high = scales::muted("red"),
  mid = "white",
  ...
)
}
\arguments{
\item{low}{color for low end of gradient}

\item{high}{color for high end of gradient}

\item{mid}{color for middle point of gradient}

\item{...}{additional arguments passed to scale_color_gradient2(...)}
}
\description{
Wrapper for scale_color_gradient2, with preferred default parameters
}
\examples{

 df.dat <- getDEG(so.query, return.list = F)
 plt.volcano <-  miko_volcano(df.deg = df.dat) + scale_color_miko()

}
