% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{scale_fill_miko}
\alias{scale_fill_miko}
\title{Gradient fill scale}
\usage{
scale_fill_miko(
  low = scales::muted("blue"),
  high = scales::muted("red"),
  mid = "white",
  ...
)
}
\arguments{
\item{low}{color for low end of gradient}

\item{high}{color for high end of gradient}

\item{mid}{color for middle point of gradient}

\item{...}{additional arguments passed to scale_color_gradient2(...)}
}
\description{
Wrapper for scale_fill_gradient2, with preferred default parameters
}
