% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{scExpression.UMAP}
\alias{scExpression.UMAP}
\title{Cell-level gene expression projected on UMAP}
\usage{
scExpression.UMAP(
  so,
  query.gene,
  x.label = "UMAP 1",
  y.label = "",
  plot.name = NULL
)
}
\arguments{
\item{so}{Seurat Object}

\item{query.gene}{Character. Gene name to plot.}

\item{x.label}{Character. X axis label.}

\item{y.label}{Character. Y axis label.}

\item{plot.name}{Character. Plot title.}
}
\value{
ggplot handle
}
\description{
UMAP plot with cell-level gene expression for queried gene.
}
