% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{scoreStem}
\alias{scoreStem}
\title{Calculate stem index}
\usage{
scoreStem(
  object,
  method = c("CCAT", "StemSC", "CytoTRACE"),
  min.pct = 0,
  assay = DefaultAssay(object),
  batch = NULL,
  verbose = T
)
}
\arguments{
\item{object}{Seurat object}

\item{method}{method used to calculate stem index. One of "CCAT", "StemSC", or "CytoTrace"}

\item{min.pct}{minimum expressing fraction. Values of 0.005-0.01 recommended if data set is large.}

\item{assay}{assay used for computing stem index. DefaultAssay(object) is taken if not specified.}

\item{batch}{meta data feature containing batch memberships. Only used for CytoTrace algorithm.}

\item{verbose}{print progress. Default is T.}
}
\value{
numeric vector
}
\description{
Calculate stem index at single-cell resolution using CCAT, StemSC or CytoTrace algorithms.
}
\examples{

ccat.score <- scoreSTEM(object = seurat.object, method = "CCAT")

}
\seealso{
\code{\link{CompCCAT}} for computing CCAT scores, \code{\link{StemSC}} for StemSC scores, and \code{\link{CytoTRACE}} for CytoTrace scores.
}
\author{
Nicholas Mikolajewicz
}
