% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{searchAnnotations}
\alias{searchAnnotations}
\title{Search Reactome/GO databases for terms that match query}
\usage{
searchAnnotations(query, db = NULL, species = NULL, ontology = NULL)
}
\arguments{
\item{query}{Query term. A character.}

\item{db}{Database to search, if specified. One of:
\itemize{
\item "Reactome"
\item "GO"
}}

\item{species}{Character specifying species, if specified. One of:
\itemize{
\item "Hs"
\item "Mm"
}}

\item{ontology}{Character specifying GO ontology to filter terms by. If unspecified, one or more of:
\itemize{
\item "BP" - biological processes
\item "MF" - molecular functions
\item "CC" - cellular compartment
}}
}
\value{
dataframe of terms that match query
}
\description{
Searches Reactome/GO databases using query stem (i.e., incomplete term) and returns all matches.
}
