% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_functions.R
\name{setResolution}
\alias{setResolution}
\title{Set cluster resolution}
\usage{
setResolution(so, cluster.resolution)
}
\arguments{
\item{so}{Seurat Object}

\item{cluster.resolution}{Seurat Object}
}
\value{
Seurat object
}
\description{
Set 'Seurat_Clusters' metadata entry to quieried cluster resolution
}
