% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{signatureCoherence}
\alias{signatureCoherence}
\title{Evaluate signature coherence.}
\usage{
signatureCoherence(
  object = NULL,
  ms.result = NULL,
  genelist,
  slot = "data",
  assay = DefaultAssay(object),
  coherence.threshold = 0.1,
  show.grid = T,
  coherent.ms = T,
  cor.method = "spearman",
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{ms.result}{Output from runMS() for provided genelist. If not specified, signatures are computed using runMS.}

\item{genelist}{Named list of genes used to compute gene signature.}

\item{slot}{Seurat slot used for gene expression correlations. Default is 'data.'}

\item{assay}{Seurat assay. Default is DefaultAssay(object).}

\item{coherence.threshold}{Coherence threshold. Default is 0.1.}

\item{show.grid}{Logical to show grid on coherence graph.}

\item{coherent.ms}{Logical to recompute coherent signature following coherence thresholding.}

\item{cor.method}{correlation metod used to determine signature coherence. Options: spearman, pearson. Default is spearman.}

\item{...}{additional parameters passed to scMiko::runMS().}
}
\value{
list of results
}
\description{
Evaluates signature coherence by computing signature score (using runMS) and then calculating correlations between signature components (i.e., gene expression) and signature score. Genes for which correlations do not exceed coherence threshold are dropped and remaining genes are used to calculate new coherent signature.
}
\examples{

}
\seealso{
\code{\link{runMS}}
}
\author{
Nicholas Mikolajewicz
}
