% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{sigScore}
\alias{sigScore}
\title{Calculate Miko score significance}
\usage{
sigScore(object, geneset, reduction = "umap")
}
\arguments{
\item{object}{Seurat Object. Must contain miko scores (i.e., must run `mikoScore` prior to `sigScore`)}

\item{geneset}{gene set list used for miko scoring (i.e., those provided as input into `mikoScore`).}

\item{reduction}{Dimensionality reduction to add to cell-level statistics data.frame. Intended for downstream use (e.g., viewing miko scores in UMAP space).}
}
\value{
list of data frames containing cell- and cluster-level statistics. Also return plot comparing raw (unstandardized) module score with Miko scores.
}
\description{
Miko scores are transformed to p values and corrected for multiple-comparisons using Benjamini Hochberg method.
}
\seealso{
\code{\link{AddSModuleScore}} for standardized module scoring, \code{\link{mikoScore}} for miko scoring
}
\author{
Nicholas Mikolajewicz
}
