% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sim2adj}
\alias{sim2adj}
\title{Compute adjaceny matrix from similary (correlation) matrix}
\usage{
sim2adj(s.mat, soft.power, network.type)
}
\arguments{
\item{s.mat}{Similarity matrix. Symmetric matrix with rows and columns corresponding to genes, and diagonal = 1.}

\item{soft.power}{Numeric specifying soft power used to calculate adjacency matrix.}

\item{network.type}{character specifying network type. Must be one of "unsigned", "signed", or "signed hybrid".}
}
\value{
adjaceny matrix (a.mat)
}
\description{
Compute adjaceny matrix from similary (correlation) matrix for different network types.
}
\examples{

# Convert similary to adjacency
a.mat <- sim2adj(s.mat, soft.power = 2, network.type = "signed")

}
