% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{snip}
\alias{snip}
\title{Winsorize values at lower and upper quantiles.}
\usage{
snip(x, lower.quantile = 0.01, upper.quantile = 0.99)
}
\arguments{
\item{x}{Numeric vector}

\item{lower.quantile}{lower quantile [0, 1]. Default = 0.01}

\item{upper.quantile}{lower quantile [0, 1]. Default = 0.99}
}
\description{
Winsorize values at lower and upper quantiles.
}
\author{
Nicholas Mikolajewicz
}
