% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{SNNExpression}
\alias{SNNExpression}
\alias{summarizeModules}
\title{Summarize SSN, ICA or NMF gene module analyses}
\usage{
summarizeModules(
  cell.object,
  gene.object,
  module.type = c("ssn", "ica", "nmf"),
  gene.list = NULL,
  reduction = NULL,
  connectivity_plot,
  group_by = "seurat_clusters",
  raster = F,
  raster.threshold = 10000,
  verbose = T
)
}
\arguments{
\item{cell.object}{Seurat object (gene x cell). Same as input to runSSN(object = cell.object, ...).}

\item{gene.object}{Seurat object (cell x gene) obtained from SSN analysis.}

\item{module.type}{Which gene modules to summarize
\itemize{
\item "ssn" - scale-free shared nearest neighbor network
\item "ica" - independent component analysis
\item "nmf" - non-negative matrix factorization
}}

\item{gene.list}{feature to visualize expression.}

\item{reduction}{specify which reduction slot use for summary. Ignored if `module.type` = "ssn".}

\item{group_by}{which meta feature to group by. Default is "seurat_clusters".}

\item{raster}{whether UMAPs are rasterized (recommended for large datasets).}

\item{raster.threshold}{cell count threshold at which to switch to rasterized plots. Default is 10000.}

\item{verbose}{Print progress. Default is T.}

\item{connecitivity.plot}{"plot_edge" generated by SSNConnectivity function. If not specified, not edges are plotted in SSN graph plot.}
}
\value{
List of summarize gene module results, including expression heatmaps and pathway enrichments
}
\description{
Summarize SSN, ICA or NMF gene module analyses. Expression heatmaps and gene module enrichments are generated to facilitate analysis and interpretation of gene modules.
}
\examples{

}
\seealso{
\code{\link{runSSN}} for SSN analysis (gene.object), \code{\link{pruneSSN}} for module features (gene.list), \code{\link{SSNConnectivity}} for connectivity plot.
}
\author{
Nicholas Mikolajewicz
}
