% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sparse2dense}
\alias{sparse2dense}
\title{Convert sparse matrix to dense matrix}
\usage{
sparse2dense(mat.sparse, block.size = 10000, transpose = F, verbose.error = F)
}
\arguments{
\item{mat.sparse}{Sparse matrix.}

\item{block.size}{If large dataset, construct dense matrix column-wise block-by-block, where block size is specified by numeric block.size parameter. Default is 10000.}

\item{transpose}{Logical specifying whether to transpose data.}

\item{verbose.error}{Logical specifying whether to print error message in case of large dataset which requires block-by-block construction.}
}
\value{
dense matrix
}
\description{
Convert sparse matrix to dense matrix. Developed to handle large datasets, by constructing dense matrix block-by-block.
}
\examples{

# Get sparse matrix
exp.mat <- so@assays[[current.assay]]@data

# convert to dense matrix
mat.dense <- sparse2dense(mat.sparse, transpose = T)

}
\seealso{
\code{\link{sparse2df}}
}
\author{
Nicholas Mikolajewicz
}
