% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{speciesConvert}
\alias{speciesConvert}
\title{Convert gene symbol representation to Hs or Mm}
\usage{
speciesConvert(query.gene, gene.list, which.species)
}
\arguments{
\item{query.gene}{Character. Query gene symbol}

\item{gene.list}{Character(s). Vector of all genes under consideration}

\item{which.species}{Character. Species}
}
\value{
Gene Symbol (character)
}
\description{
Convert gene symbol to approriate species (as specified), while ensuring that gene is present in reference gene list.
}
