% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{SSNConnectivity}
\alias{SSNConnectivity}
\title{SSN connectivity plot}
\usage{
SSNConnectivity(
  gene.object,
  gene.list = NULL,
  quantile_threshold = 0.9,
  raster_dpi = 200,
  edge.alpha = 0.015,
  edge.color = "black",
  node.color = "grey",
  node.alpha = 1,
  node.weights = F,
  node.size.min = 1,
  node.size.max = 5,
  do.label = T,
  label.size = 5,
  verbose = T
)
}
\arguments{
\item{gene.object}{Seurat object (cell x gene) obtained from SSN analysis.}

\item{gene.list}{named list of gene program features/genes. If omitted, feature-annotated connectivity plot is not generated.}

\item{quantile_threshold}{Quantile threshold for visualized SSN graph edges. Default is 0.9.}

\item{raster_dpi}{prefix added to each named entry in list. Default is "".}

\item{edge.alpha}{alpha [0,1] parameter (i.e., transparency) for network edges. Default is 0.015. Use larger values for less dense networks.}

\item{edge.color}{color of network edges. Default is 'black'.}

\item{node.color}{color of node edges. Default is 'grey'.}

\item{node.alpha}{alpha [0,1] parameter for nodes. Default is 1 (obaque).}

\item{node.weights}{scale node size by connectivity weights. Default is F.}

\item{node.size.min}{minimal node size. Ignored if node.weights = F. Default is 1.}

\item{node.size.max}{maximal node size. Ignored if node.weights = F. Default is 5.}

\item{do.label}{Show gene program IDs  on network graph. Default is T.}

\item{label.size}{Size of gene program IDs on network graph. Default is 5.}

\item{verbose}{Print progress. Default is T.}
}
\value{
Returns 2 variants of SSN connectivity plot along with data.frame used to generate plots.
}
\description{
generates SSN connectivity plot used to visualize transcriptomic gene network.
}
\references{
\url{https://nmikolajewicz.github.io/scMiko/articles/Module_Detection.html}
}
\seealso{
\code{\link{runSSN}} for SSN analysis (gene.object), \code{\link{pruneSSN}} for gene program features (gene.list)
}
\author{
Nicholas Mikolajewicz
}
