% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{SSNExpression}
\alias{SSNExpression}
\title{SSN connectivity plot}
\usage{
SSNExpression(
  cell.object,
  gene.object,
  group_by = "seurat_clusters",
  features = NULL,
  max_cells = 20000,
  connectivity.plot = NULL
)
}
\arguments{
\item{cell.object}{Seurat object (gene x cell). Same as input to runSSN(object = cell.object, ...).}

\item{gene.object}{Seurat object (cell x gene) obtained from SSN analysis.}

\item{group_by}{meta feature used for grouping. Must be present in `cell.object`.}

\item{features}{feature to visualize expression.}

\item{max_cells}{maximum number of cells used. Default is 20000. If number of cells in `cell.object` exceeds `max_cells`, `cell.object` is downsampled.}

\item{connecitivity.plot}{"plot_edge" generated by SSNConnectivity function. If not specified, not edges are plotted in SSN graph plot.}
}
\value{
Returns 2 variants of SSN connectivity plot along with data.frame used to generate plots.
}
\description{
generates SSN connectivity plot used to visualize transcriptomic gene network.
}
\references{
\url{https://nmikolajewicz.github.io/scMiko/articles/Module_Detection.html}
}
\seealso{
\code{\link{runSSN}} for SSN analysis (gene.object), \code{\link{pruneSSN}} for gene program features (gene.list), \code{\link{SSNConnectivity}} for connectivity plot.
}
\author{
Nicholas Mikolajewicz
}
