% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{standardizeNonParametric}
\alias{standardizeNonParametric}
\title{Non-parametric standardization of matrix}
\usage{
standardizeNonParametric(z, which.dim = 1)
}
\arguments{
\item{z}{matrix}

\item{which.dim}{dimension. Default is 1 (rows). Use 2 for columns.}
}
\description{
Center using median and scale using median absolute deviations
}
\author{
Nicholas Mikolajewicz
}
