% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{subsetDimRed}
\alias{subsetDimRed}
\title{Get dimensional reduction from Seurat Object for subset of data}
\usage{
subsetDimRed(
  so,
  which.features,
  groups = "seurat_clusters",
  reduction = "umap"
)
}
\arguments{
\item{so}{Seurat Object}

\item{which.features}{Character vector of features to subset on.}

\item{reduction}{A character specifying which reduction to retrieve from Seurat Object. Default is 'umap'.}
}
\value{
list of results
}
\description{
Get dimensional reduction from Seurat Object for subset of data.
}
\author{
Nicholas Mikolajewicz
}
