% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{subsetSeurat}
\alias{subsetSeurat}
\title{Subset Seurat Object}
\usage{
subsetSeurat(object, subset.df)
}
\arguments{
\item{object}{Seurat object}

\item{subset.df}{Data.frame specifying which field (subset.df$field) to subset on, and which field entries to retain (subset.df$subgroups).}
}
\value{
Seurat object
}
\description{
Subset Seurat object according to specified meta data field. Only specified meta data entries are retained, while remaining of data is omitted.
}
\examples{

# define subset parameters
subset.df <- data.frame(field = "seurat_clusters", subgroups = c(0,2,3,5,6,7,9,15)) # tumor population from Renca 4000ds T12 sample

# subset data
so <- subsetSeurat(so, subset.df)

}
