% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg_functions.R
\name{summarizeHG}
\alias{summarizeHG}
\title{Summarize hypergeometric enrichment results}
\usage{
summarizeHG(
  hg.res,
  fdr.filter = 1,
  do.plot = T,
  show.n = 5,
  genesets = NULL,
  str.wrap.width = 25,
  col.pal = NULL,
  plt.title = "",
  pathway.name.size = 7
)
}
\arguments{
\item{hg.res}{RunHG output list}

\item{fdr.filter}{FDR filter threshold. Default 1 (i.e., include all)}

\item{do.plot}{Logical to generate enrichment plots for each geneset.}

\item{show.n}{Number of top enrichments to visualize in plots. Ignored if do.plot == F.}

\item{genesets}{Geneset list used for HG enrichment (i.e., RunHG input). Optional.}

\item{str.wrap.width}{positive integer giving target line width in characters. A width less than or equal to 1 will put each word on its own line. Ignored if do.plot == F.}

\item{plt.title}{Character specifying plot title. Ignored if do.plot == F.}

\item{pathway.name.size}{Font size for pathway names in plots. Ignored if do.plot == F.}
}
\value{
list of summarized results
}
\description{
Summarize hypergeometric enrichment results. Collapses HG result list into dataframe, and generates plots to visualize enrichments for each geneset.
}
\author{
Nicholas Mikolajewicz
}
\concept{enrichment}
