% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sym2ens}
\alias{sym2ens}
\title{Convert gene symbol to ensembl}
\usage{
sym2ens(my.symbols, my.species)
}
\arguments{
\item{my.symbols}{Character. Vector of Gene symbols.}

\item{my.species}{Character. Species.}
}
\value{
data.frame mapping gene Symbols to Ensembl
}
\description{
Gene symbol is converted to ensembl id using org.Hs.eg.db or org.Mm.eg.db annotation databases.
}
