% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sym2entrez}
\alias{sym2entrez}
\title{Convert gene symbol to entrez}
\usage{
sym2entrez(my.symbols, my.species)
}
\arguments{
\item{my.symbols}{Character. Vector of Gene symbols.}

\item{my.species}{Character. Species.}
}
\value{
data.frame mapping gene Symbols to Entrez
}
\description{
Gene symbol is converted to entrez id using org.Hs.eg.db or org.Mm.eg.db annotation databases.
}
\author{
Nicholas Mikolajewicz
}
