% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{term2id}
\alias{term2id}
\title{Map Reactome/GO term to ID}
\usage{
term2id(term, db = "Reactome", species = NULL)
}
\arguments{
\item{term}{Reactome/GO term}

\item{db}{Database to use for mapping. One of:
\itemize{
\item "Reactome" - Default
\item "GO"
}}

\item{species}{Character specifying species. If specified, one of:
\itemize{
\item "Hs"
\item "Mm"
}}
}
\value{
Reactome/GO ID
}
\description{
Map Reactome/GO term to ID
}
\examples{

# map GO term to GO id
id <- term2id(term, db = "GO", species = "Hs")

# get geneset for specified GO id (returns character of genes)
gene.set <- id2geneset(id, "Hs")

}
