% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{theme_miko}
\alias{theme_miko}
\title{scMiko Theme}
\usage{
theme_miko(style = "bw", legend = F, grid = F, bold.title = T)
}
\arguments{
\item{style}{Choise of 'theme_classic' or 'theme_bw' called using 'classic' or 'bw', respectively. Default is 'bw'}

\item{legend}{Logical to include legend. Default is F.}

\item{grid}{Logical to include grid. Default is F.}

\item{bold.title}{Logical to bold title. Default is T.}

\item{viridis.color}{Logical to use virdis color pallete for 'color'. Default is F.}

\item{viridis.fill}{Logical to use virdis fill pallete for 'fill'. Default is F.}
}
\value{
ggplot2 theme object
}
\description{
Custom themes applied to ggplot2-based plots.
}
\examples{


}
