% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ulength}
\alias{ulength}
\title{Number of unique values}
\usage{
ulength(x)
}
\arguments{
\item{x}{object}
}
\value{
numeric
}
\description{
Number of unique values.
}
\examples{

x <- c(1,1,1,1,2,2,2,3,)
n.x <- ulength(x)

}
\author{
Nicholas Mikolajewicz
}
