% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{uniqueFeatures}
\alias{uniqueFeatures}
\title{Get unique features from metadata column in seurat object.}
\usage{
uniqueFeatures(object, group, is.numeric = F)
}
\arguments{
\item{object}{Seurat object}

\item{group}{name of metadata column to get features for.}

\item{is.numeric}{Logical to assert numerical class and order features.}
}
\value{
Seurat object
}
\description{
Get unique features from metadata column in seurat object.
}
\author{
Nicholas Mikolajewicz
}
