% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{updateCentralLog}
\alias{updateCentralLog}
\title{Update central log}
\usage{
updateCentralLog(
  Module,
  clog.file = "moduleLog.csv",
  log.path = if (exists("data.path")) data.path else "",
  user.id = if (exists("user")) user else "guest",
  run.notes = NA,
  pdf.flag = save.pdf
)
}
\arguments{
\item{Module}{Module Name/ID. e.g., "M01" is used to specify Module 1.}

\item{clog.file}{Central log file name. A character.}

\item{log.path}{Path to central log file. Set to data.path specified in .RProfile by default.}

\item{user.id}{User name. Set to user specified in .RProfile by default.}

\item{run.notes}{Additional notes to describe current run.}

\item{pdf.flag}{Logical indicating whether PDF figures were generated.}
}
\description{
Update central log. Central log keeps track of every scPipeline run and includes information about the module run, data used, user, date.
}
\examples{

# update central log (usually at the end of a successful module run)
updateCentralLog(Module = "M01", pdf.flag = F)

}
\author{
Nicholas Mikolajewicz
}
