% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{updateDimNames}
\alias{updateDimNames}
\title{Ensure that all dimNames are correctly specified in Seurat Object}
\usage{
updateDimNames(so)
}
\arguments{
\item{so}{Seurat Object}
}
\value{
Seurat Object
}
\description{
Ensure that all dimNames are correctly specified in Seurat Object. If incorrectly specified, subsetting functions will lead to unexpected results.
}
\examples{

so <- UpdateDimNames(so)

}
\author{
Nicholas Mikolajewicz
}
