% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{updateGeneSets}
\alias{updateGeneSets}
\title{Create/update list of genessets}
\usage{
updateGeneSets(input.file, output.file, dir = "")
}
\arguments{
\item{input.file}{Excel file (input). Must have ".xlsx" suffix. A character.}

\item{output.file}{Rdata file (output). Must have ".Rdata" suffix. A character.}

\item{dir}{Directory of input and output file (same folder). A character.}
}
\value{
List of data.frames saved as Rdata file.
}
\description{
Takes genessets stored in Excel sheets, and converts them to dataframes stored in lists.
}
