% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{upset.Plot}
\alias{upset.Plot}
\title{Upset plot}
\usage{
upset.Plot(gene.sets, row.title = "", column.title = "")
}
\arguments{
\item{gene.sets}{named list of genesets, where names specify name of gene set, and entries are character vectors specifying genes belongs to the respective set.}

\item{row.title}{Row title.}

\item{column.title}{Column title.}
}
\value{
plot handle
}
\description{
Generate upset plot, comparing intersection between different (gene)sets. Uses ComplexHeatmap Package.
}
\examples{

Generate upset plot for list of genesets
plt.upset <- upset.Plot(gene.sets, row.title = "Genesets")

# print plot
print(plt.upset)

}
