% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{value2col}
\alias{value2col}
\title{Convert values to color gradient}
\usage{
value2col(
  values,
  limit = NULL,
  gradient.length = 100,
  low.col = "skyblue",
  mid.col = "grey",
  high.col = "tomato"
)
}
\arguments{
\item{limit}{numeric specifying limit of color range. If unspecified, limit <- max(c(abs(min(Values)), abs(max(Values))))}

\item{gradient.length}{Numeric specifying number of bins to split gradient into. Default is 100.}

\item{low.col}{Color representing low values. Default is "skyblue"}

\item{mid.col}{Color representing mid values. Default is "grey"}

\item{high.col}{Color representing high values. Default is "tomato"}

\item{Values}{vector of numerical values to convert to colors}
}
\value{
vector of colors.
}
\description{
Converts ranges of values to corresponding color along color gradient, specified by 3 colors (low color, middle color and high color)
}
\examples{

# get edge colors
col.lim <- max(c(abs(min(vis.links$value)), abs(max(vis.links$value))))
edge.colors <- value2col(vis.links$value, limit = col.lim)

}
\seealso{
\code{\link{colorRampPalette}}
}
