% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{varExpPCA}
\alias{varExpPCA}
\title{Get optimal number of principal components}
\usage{
varExpPCA(object, variance_explained = 0.9, reduction_name = "pca")
}
\arguments{
\item{object}{Seurat object}

\item{variance_explained}{proportion of variance explained. Default is 0.9.}

\item{reduction_name}{name of slot with PCA reduction. default is "pca".}
}
\value{
number of PCs explaining specified amount of variance
}
\description{
Get number of PCs that explain a pre-specified amount of variance.
}
\author{
Nicholas Mikolajewicz
}
