% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{variableGenes.Plot}
\alias{variableGenes.Plot}
\title{Plot variable genes in Seurat Object}
\usage{
variableGenes.Plot(
  so,
  gNames,
  set_name = NULL,
  top.n.genes = 10,
  pt.colors = c("black", "tomato"),
  ...
)
}
\arguments{
\item{so}{Seurat Object}

\item{gNames}{Named gene list; entries are Symbols, names are Ensemble.}

\item{set_name}{Character specfiying name of dataset. Optional.}

\item{top.n.genes}{Numeric. Top n genes to label on plot.}

\item{pt.colors}{colors to specify non-variable/variable status.}

\item{...}{additional arguments passed to Seurat::VariableFeaturePlot(...)}
}
\value{
ggplot handle
}
\description{
Using Seurat's VariableFeaturePlot, variable genes are identified and plotted. Top N variable genes are labeled.
}
