% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{varyingPaths.Tempora}
\alias{varyingPaths.Tempora}
\title{Identify temporally varying pathways using Tempora}
\usage{
varyingPaths.Tempora(
  object,
  pval_threshold = 0.05,
  adjust.p = T,
  pathway.filter = 3
)
}
\arguments{
\item{object}{Tempora Object}

\item{pval_threshold}{P-value threshold to determine the significance of pathway enrichment over time. Default is 0.05.}

\item{adjust.p}{Flag to adjust p value using BH correction. Default is TRUE.}

\item{pathway.filter}{Numeric specifying z score threshold used to pre-filter pathways. Default is 3.}
}
\value{
List containing generalized additive model (GAMs) fits, plots and p-values
}
\description{
Identify temporally varying pathways using Tempora
}
\examples{

}
\author{
Gary Bader et. al
}
