% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{vd_Inputs}
\alias{vd_Inputs}
\title{Specify inputs for variance decomposition analysis}
\usage{
vd_Inputs(
  object,
  vd_model.list,
  features = NULL,
  pct.min = 0,
  variable.features = F,
  subsample.factor = 1
)
}
\arguments{
\item{object}{Seurat object.}

\item{vd_model.list}{Output from vd_Formula.}

\item{pct.min}{Minimal expressing fraction for genes to be included in analysis. Default is 0.}

\item{variable.features}{Logical specifying whether to use variable features only. If true, looks for variable features within provided Seurat object.}

\item{subsample.factor}{Numeric [0,1] specfying how to subsample (i.e., downsample) data. Default is 1 (no subsampling)}

\item{Features}{Features to include in analysis. If specified, pct.min and variable features are ignored.}
}
\value{
list of inputs for vd_Run() function.
}
\description{
Step 2 of variance decomposition analysis (see examples). Given Seurat object and vd_Formula output, input list for variance decomposition are generated.
}
\examples{

parameter.list <- list(
 covariates = c( "cluster", "percent.mt", "batch", "cycle", "seq.coverage"),
 interactions = c("batch:cluster")
)

# step 1: model formulation
vd_model.list <- vd_Formula(object = so.query,
                           covariates = parameter.list$covariates,
                           interactions = parameter.list$interactions)

# step 2: prep model inputs
vd_inputs.list <- vd_Inputs(object = so.query, vd_model.list = vd_model.list, features = NULL,
                           pct.min =  0.9, variable.features = F, subsample.factor = 1)

# step 3: run variance decomposition
vd_results.list <- vd_Run(vd_inputs.list, n.workers = 20)

# step 4 (optional): visualize UMAP distribution of covariates
plt.umap.list <- vd_UMAP(object = so.query, vd_model.list = vd_model.list)

# step 5 (optional): visualize decomposition
res.var2 <- vd_results.list$varPart.format1
plt.var <- plotVarPart( res.var2 ) +
 theme_miko() +
 labs(title = "Variance Decomposition", subtitle = "Linear Mixed-Effects Model")  +
 theme(axis.text.x = element_text(angle = 45, hjust = 1))

}
\seealso{
\code{\link{vd_Run}}
}
\author{
Nicholas Mikolajewicz
}
\concept{vd}
