% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{vd_Run}
\alias{vd_Run}
\title{Perform Variance Decomposition Analysis}
\usage{
vd_Run(vd_inputs.list, n.workers = 20)
}
\arguments{
\item{vd_inputs.list}{Output from scMiko::vd_Input() function.}

\item{n.workers}{Number of workers to use (for parallel implementation; uses foreach package)}
}
\value{
List of results summarizing variance explained by each model covariate.
}
\description{
Step 3 of variance decomposition analysis (see examples).
}
\examples{

parameter.list <- list(
 covariates = c( "cluster", "percent.mt", "batch", "cycle", "seq.coverage"),
 interactions = c("batch:cluster")
)

# step 1: model formulation
vd_model.list <- vd_Formula(object = so.query,
                           covariates = parameter.list$covariates,
                           interactions = parameter.list$interactions)

# step 2: prep model inputs
vd_inputs.list <- vd_Inputs(object = so.query, vd_model.list = vd_model.list, features = NULL,
                           pct.min =  0.9, variable.features = F, subsample.factor = 1)

# step 3: run variance decomposition
vd_results.list <- vd_Run(vd_inputs.list, n.workers = 20)

# step 4 (optional): visualize UMAP distribution of covariates
plt.umap.list <- vd_UMAP(object = so.query, vd_model.list = vd_model.list)

# step 5 (optional): visualize decomposition
res.var2 <- vd_results.list$varPart.format1
plt.var <- plotVarPart( res.var2 ) +
 theme_miko() +
 labs(title = "Variance Decomposition", subtitle = "Linear Mixed-Effects Model")  +
 theme(axis.text.x = element_text(angle = 45, hjust = 1))

}
\seealso{
\code{\link{vd_Input}}
}
\author{
Nicholas Mikolajewicz
}
