% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{wideDF2namedList}
\alias{wideDF2namedList}
\title{Convert wide data.frame to named list}
\usage{
wideDF2namedList(df.wide)
}
\arguments{
\item{df.wide}{wide data.frame}
}
\value{
named list
}
\description{
Convert wide data.frame to named list. Column entries are used as list entries, and each entry is named using the corresponding column name in the data.frame.
}
\examples{

# get wide data frame
CancerSEA_Hs <-geneSets[["CancerSEA_Hs"]]

# convert wide data frame to named list
my.list <- namedList2longDF(CancerSEA_Hs)

}
\seealso{
\code{\link{namedList2wideDF}}
}
\author{
Nicholas Mikolajewicz
}
