% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn_functions.R
\name{wnn_Components}
\alias{wnn_Components}
\title{Compute network component UMAPs and visualize component weights.}
\usage{
wnn_Components(
  object,
  exp.list,
  dim.lists,
  plt.wnn.umap,
  umap.min.dist = 0.1,
  umap.n.neighbors = 20
)
}
\arguments{
\item{object}{wnn integrated seurat object (wnn_Run output)}

\item{exp.list}{List of expression matrices (wnn_Run list input)}

\item{dim.lists}{List of PC numbers (wnn_Run output)}

\item{plt.wnn.umap}{ggplot handle for integrated network umap (wnn_Run output)}

\item{umap.min.dist}{min.dist for umap embedding. See RunUMAP(). Default: 0.1}

\item{umap.n.neighbors}{n.neighbor for umap embedding. See RunUMAP(). Default: 20}
}
\value{
list of ggplot handle and object
}
\description{
Compute network component UMAPs and visualize component weights.
}
\author{
Nicholas Mikolajewicz
}
