% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn_functions.R
\name{wnn_Run}
\alias{wnn_Run}
\title{Run WNN Multi-Modal Integration.}
\usage{
wnn_Run(
  object,
  wnn.knn = 20,
  umap.knn = 20,
  umap.min.dist = 0.1,
  do.scale = F,
  do.center = F,
  pca.thres = 0.9,
  cres = 1,
  min.pct = 0.25,
  split.var = "seurat_clusters"
)
}
\arguments{
\item{object}{Seurat object or list of expression matrices. If seurat object, expression matrices are extracted. If list, assumes that expression matrix entries have column-wise genes and row-wise cells.}

\item{wnn.knn}{the number of multimodal neighbors to compute. 20 by default}

\item{umap.knn}{This determines the number of neighboring points used in local approximations of manifold structure. Larger values will result in more global structure being preserved at the loss of detailed local structure. In general this parameter should often be in the range 5 to 50. Default: 20}

\item{umap.min.dist}{This controls how tightly the embedding is allowed compress points together. Larger values ensure embedded points are moreevenly distributed, while smaller values allow the algorithm to optimise more accurately with regard to local structure. Sensible values are in the range 0.001 to 0.5. Default: 0.1}

\item{do.scale}{Logical to scale expression. Default is F.}

\item{do.center}{Logical to center expression. Default is F.}

\item{pca.thres}{Variance explained threshold for PC component inclusion. Default is 0.9.}

\item{cres}{Cluster resolution for integrated network. Default is 1.}

\item{min.pct}{Minimum expression fraction for inclusion in network integration. Default is 0.25. Ignored if object is list.}

\item{split.var}{Grouping variable for expression fraction filter. Default is 'seurat_clusters'. Ignored if object is list.}

\item{normalize.margin}{If specifies, normalize across rows/cells (1) or columns/genes (2)}
}
\value{
list of integrated results
}
\description{
Run WNN Multi-Modal Integration. Modified wrapper for seurat WNN workflow.
}
\author{
Nicholas Mikolajewicz
}
