% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{z2p}
\alias{z2p}
\title{Convert Z score to p value}
\usage{
z2p(z, two.sided = T)
}
\arguments{
\item{z}{Z score}

\item{two.sided}{Specify where two-sided distribution is used for p value calculation. Default is T.}
}
\value{
p value
}
\description{
Convert Z score to p value
}
\seealso{
\code{\link{pnorm}}
}
