% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{avgGroupExpression}
\alias{avgGroupExpression}
\title{Get average group expression in Seurat object}
\usage{
avgGroupExpression(so, which.data = "data", which.group = "seurat_clusters")
}
\arguments{
\item{so}{Seurat Object}

\item{which.data}{Character specfying which data slot. Default is "data".}

\item{which.group}{Character specfying group field in Seurat metadata. Default is "seurat_clusters".}
}
\value{
Numeric vector, ordered
}
\description{
Get average group expression in Seurat object
}
