% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cleanFilterGenes}
\alias{cleanFilterGenes}
\title{Clean and filter gene list}
\usage{
cleanFilterGenes(genes, so, which.species)
}
\arguments{
\item{genes}{Character vector of genes}

\item{so}{Seurat Object}

\item{which.species}{Species}
}
\value{
Character vector of genes
}
\description{
Clean gene list (i.e., remove NAs), convert to appropraite species, and keep only those present in seurat object.
}
