% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{detectSpecies}
\alias{detectSpecies}
\title{Determine species based on gene representation}
\usage{
detectSpecies(object)
}
\arguments{
\item{object}{Seurat object, gene expression matrix (rownames are genes), or gene character vector.}
}
\value{
Species (Hs or Mm)
}
\description{
Determine species (Hs, Mm) based on gene (symbol, ensembl) representation
}
\examples{

my.species <- detectSpecies(object)

}
\author{
Nicholas Mikolajewicz
}
