% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{downsampleSeurat}
\alias{downsampleSeurat}
\title{Downsample single cell data}
\usage{
downsampleSeurat(so, subsample.factor)
}
\arguments{
\item{so}{Seurat Object}

\item{subsample.factor}{Numeric [0,1]. Factor to downsample data by.}
}
\value{
Seurat Object
}
\description{
Downsample data in Seurat object by specified factor
}
