% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{enrichGO.fisher}
\alias{enrichGO.fisher}
\title{Over-representation enrichment of GO terms using weighted fisher method}
\usage{
enrichGO.fisher(
  gene.candidates,
  gene.universe,
  which.species,
  which.ontology = "BP",
  p.threshold = 0.001,
  padj.threshold = 0.05,
  topGO.object = NULL
)
}
\arguments{
\item{gene.candidates}{Vector of candidate genes.}

\item{gene.universe}{Vector of all genes.}

\item{which.species}{Character specifying species. "Mm" or "Hs".}

\item{which.ontology}{Character specifying ontology. "BP", "MF", or "CC".}

\item{p.threshold}{p value threshold. Default is 0.001.}

\item{padj.threshold}{adjusted p value threhsold (BH). Default is 0.05.}

\item{topGO.object}{topGO object (Optional). If unspecified, new topGO object is new created. If specified, gene list is used to updated existing object.}
}
\value{
list of 2 data.frames (unadjusted.results (p<0.001) and adjusted.results (padj < 0.05)) and a topGO object.
}
\description{
Over-representation enrichment of GO terms using weighted fisher method
}
