% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ens2sym.so}
\alias{ens2sym.so}
\title{Convert gene ensemble to symbol in seurat object}
\usage{
ens2sym.so(so, gNames.list, convert.RNA = TRUE)
}
\arguments{
\item{reference.genes}{Named vector of genes; names are ENSEMBL, entries are SYMBOL.}

\item{query.genes}{vector of query genes to check representation}
}
\value{
Seurat object with specified gene representation.
}
\description{
Process Seurat object so that genes in Ensemble format are converted to Symbol representation.
}
