% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{expression.Plot}
\alias{expression.Plot}
\title{Violin plot of single cell gene expression}
\usage{
expression.Plot(
  so,
  which.gene,
  e.mat = NULL,
  f.mat = NULL,
  which.group = "seurat_clusters",
  which.data = "data",
  which.assay = DefaultAssay(so),
  x.label = NULL,
  x.label.angle = NULL
)
}
\arguments{
\item{so}{Seurat object}

\item{which.gene}{query gene. Must be available in Seurat object.}

\item{e.mat}{expression matrix (gene x cell). If not provided, computed from seurat object (more computationally heavy).}

\item{f.mat}{expression fraction matrix (gene x cell). If not provided, computed from seurat object (more computationally heavy).}

\item{which.group}{Group to cluster gene expression by. Must be column header in so meta data. Default is "seurat_clusters".}

\item{which.data}{Specifies which seurat data used to compute e.mat and f.mat. Must be either "data" (default) or "scale". Used only if e.mat and f.mat are not provided.}

\item{which.assay}{Specifies which seurat assay to used to compute e.mat. Used only if e.mat is not provided. Default is DefaultAssay(so)}

\item{x.label}{x axis title (a character). Default is which.group.}

\item{x.label.angle}{rotation angle for x axis title (a numeric). Default is NULL.}
}
\value{
ggplot object
}
\description{
Draws violin plot of single cell data, stratified by specified group. Hierarchically-clustered dendrogram is overlaid to illustrate relationships between groups.
}
\examples{

# get expression plot of GPNMB stratified by cluster group.
plt.sgExp.clust <- expression.Plot.dev(so.query, which.gene = "GPNMB", which.group = "seurat_clusters", which.data = "data")
print(plt.sgExp.clust)


}
