% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{expression.Plot}
\alias{expression.Plot}
\title{Violin plot of single cell gene expression}
\usage{
expression.Plot(
  so,
  which.gene,
  which.group = "seurat_clusters",
  which.data = "data"
)
}
\arguments{
\item{so}{Seurat object}

\item{which.gene}{query gene. Must be available in Seurat object.}

\item{which.group}{Group to cluster gene expression by. Must be column header in so meta data. Default is "seurat_clusters".}

\item{which.data}{Specifies which data to use for plot. Must be either "data" (default) or "scale".}
}
\value{
ggplot object
}
\description{
Draws violin plot of single cell data, stratified by specified group. Hierarchially-clustered dendrogram is overlaid to illustrate relationships between groups.
}
\examples{

# get expression plot of GPNMB stratified by cluster group.
plt.sgExp.clust <- expression.Plot.dev(so.query, which.gene = "GPNMB", which.group = "seurat_clusters", which.data = "data")
print(plt.sgExp.clust)


}
