% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{findNetworkFeatures}
\alias{findNetworkFeatures}
\title{Identify features for gene program discovery.}
\usage{
findNetworkFeatures(
  object,
  method = c("expr", "hvg", "deviance"),
  n_features = 2000,
  min_pct = 0.5,
  split_var = "seurat_clusters",
  batch_feature = NULL,
  verbose = T
)
}
\arguments{
\item{object}{Seurat object}

\item{method}{Feature selection method.
\itemize{
\item "expr" - Top expressed genes within `split_var` groups, together with variable genes.
\item "hvg" - Highly-variable genes (number specified by `n_features`)
\item "deviance" - Deviant genes (number specified by `n_features`)
}}

\item{n_features}{Number of features to use when calculating variable or deviant features. Default is 2000.}

\item{min_pct}{Minimum expression within spilt_var. Ignored if `method` is not "expr". Default is 0.5.}

\item{split_var}{Grouping variable to enforce `min_pct` criteria within. Default is 'seurat_clusters'.}

\item{batch_feature}{Variables to regress out. Ignored if `method` is not "deviance". Default is NULL.}

\item{verbose}{Show progress. Default is T.}
}
\value{
vector of features
}
\description{
Identify features for gene program discovery.
}
\references{
\url{https://nmikolajewicz.github.io/scMiko/articles/Module_Detection.html}
}
\seealso{
\code{\link{runSSN}}
}
\author{
Nicholas Mikolajewicz
}
