% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{firstup}
\alias{firstup}
\title{Uppercase first letter and lowercase rest}
\usage{
firstup(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character vector
}
\description{
Function used to naively convert gene list to mouse representation; i.e., Convert Hs to Mm gene symbols.
}
\seealso{
\code{\link{toupper}}
}
