% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_functions.R
\name{flex.asDT}
\alias{flex.asDT}
\title{Outputs datatable with print button options}
\usage{
flex.asDT(df, page_length = 50, ...)
}
\arguments{
\item{df}{data.frame}

\item{page_length}{number of entries shown per page. Default is 50.}

\item{...}{additional parameters passed to option list}
}
\value{
data.table
}
\description{
Outputs datatable with print button options
}
