% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_functions.R
\name{flex.multiTabPlot}
\alias{flex.multiTabPlot}
\alias{flex.asDT}
\title{Generate multi-tab ggplot handle list for flexdashboard}
\usage{
flex.asDT(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.table
}
\description{
Prepares list of ggplot handles for multi-tab plot presentation in flexdashboards.
}
