% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{geneSets}
\alias{geneSets}
\title{List of scMiko gene sets}
\format{
A list of data.frames - each data.frame is an independent dataset.
}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(geneSets)
}
\description{
Reference gene sets consolidated from numerous sources to faciliate cell annotation and module scoring.
}
\examples{
data(geneSets)
CancerSEA_Hs <- geneSets[["CancerSEA_Hs"]]

}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{annotation,}
\keyword{cell}
\keyword{genesets,}
\keyword{module}
\keyword{scoring}
