% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getAnnotationPathways}
\alias{getAnnotationPathways}
\title{Returns list of annotations for given Entrez gene IDs}
\usage{
getAnnotationPathways(
  query.genes,
  db = c("Reactome"),
  ontology = c("BP"),
  species = c("Hs")
)
}
\arguments{
\item{query.genes}{Entrez IDs of query genes}

\item{db}{Database to retrieve annotations from. One of:
\itemize{
\item "Reactome" - Default
\item "GO"
}}

\item{ontology}{GO ontologies to retrieve if GO db is selected. One of:
\itemize{
\item "BP" - Default. Biological processes
\item "MF" - Molecular functions
\item "CC" - Cellular components
}}

\item{species}{"Mm" or "Hs". Default is "Hs".}
}
\value{
Named list of vectors with gene sets.
}
\description{
Returns list of Reactome or GO annotations for given Entrez gene IDs
}
