% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{getDEG}
\alias{getDEG}
\title{Get differentially expressed genes}
\usage{
getDEG(
  object,
  assay = "SCT",
  data = "data",
  group_by = "seurat_clusters",
  auc.thresh = 0.6,
  fdr.thresh = 0.01,
  logFC.thresh = NA,
  pct.dif.thresh = NA,
  return.list = T,
  return.all = F,
  verbose = T
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{assay. Default "SCT".}

\item{data}{data slot. Default "data".}

\item{group_by}{Name of grouping variable (must be present in object's meta.data)}

\item{auc.thresh}{AUC threshold. Default = 0.6.}

\item{fdr.thresh}{FDR threshold. Default = 0.01.}

\item{logFC.thresh}{logFC threshold. Default = NA.}

\item{pct.dif.thresh}{Difference in expression percentage. Default = 0.}

\item{return.list}{If TRUE, return list of differentially expressed genes. If FALSE, returns table with statistics from differential expression analysis.}

\item{return.all}{If TRUE, all thresholding filters are ignored, and all results are returned.}

\item{verbose}{Print progress. Default is TRUE.}
}
\value{
data.frame or list
}
\description{
Differential expression analysis wrapper for presto::wilcoxauc(). Option to return differentially-expressed gene list (return.list = T) or statistics only (return.list = F)
}
\examples{

}
\seealso{
\code{\link{wilcoxauc}}
}
\author{
Nicholas Mikolajewicz
}
