% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getExpressionMatrix}
\alias{getExpressionMatrix}
\title{Get expression matrix from Seurat Object}
\usage{
getExpressionMatrix(
  so,
  only.variable = F,
  which.assay = NULL,
  which.data = "scale",
  use.additional.genes = NA
)
}
\arguments{
\item{so}{Seurat Object}

\item{only.variable}{Logical indicating whether to include variable features only or not.}

\item{which.assay}{Seurat assay to get data frome. Default is DefaultAssay(so).}

\item{which.data}{Specify which data to use (refers to slots in Seurat object assay). One of:
\itemize{
\item "scale" - Default
\item "data"
}}

\item{use.additional.genes}{Character vector of additional genes to include (in addition to varibale, if variable flag is specificed). Default is NA.}
}
\value{
gene x cell expression matrix
}
\description{
Get expression matrix from Seurat Object
}
\author{
Nicholas Mikolajewicz
}
