% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_functions.R
\name{importSeurat}
\alias{importSeurat}
\title{import Seurat}
\usage{
importSeurat(file, directory = NULL)
}
\arguments{
\item{file}{Seurat object file name}

\item{directory}{directory in which Seurat object is stored}
}
\value{
Seurat object
}
\description{
Import Seurat Object and preprocess using fixBarcodeLabel() and UpdateSeuratObject() Functions.
}
