% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{inferState}
\alias{inferState}
\title{Infer activation state using Gaussian decomposition}
\usage{
inferState(score, group = NULL, diffvar = TRUE, k = 5, verbose = T)
}
\arguments{
\item{score}{activity vector  from which activation state will be inferred. Must be continuous data.}

\item{group}{grouping vector. Defaultis NULL.}

\item{diffvar}{specify whether data is heteroscedastic. Default is T.}

\item{k}{Number of components to evaluate.}

\item{verbose}{Print progress. Default is TRUE.}
}
\value{
list of results
}
\description{
Infer activation state using Gaussian decomposition
}
\seealso{
\code{\link{Mclust}}
}
