% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_functions.R
\name{loadMat}
\alias{loadMat}
\title{Load gene x cell count matrix into seurat object}
\usage{
loadMat(file, dir)
}
\arguments{
\item{file}{Matrix file name.}

\item{dir}{Character. folder containing import_set file}
}
\value{
list containing Seurat Object and named gene vector.
}
\description{
Load gene x cell count matrix into seurat object. For intended performance, first column of spreadsheet must contain gene names, and first row of spreadsheet must contain cell names.
}
