% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{miko_volcano}
\alias{miko_volcano}
\title{Draw volcano plot to visualize differential expression.}
\usage{
miko_volcano(
  df.deg,
  group = NULL,
  show.n = 10,
  features = NULL,
  fdr.threshold = 0.05,
  label.size = NA
)
}
\arguments{
\item{df.deg}{Differential expression data. Dataframe output from presto::wilcoxauc() or scMiko::getDEG(..., return.list = F).}

\item{group}{group to visaulize data for. Must be entry in 'group' column of df.deg.}

\item{show.n}{Top n features to label. Default is 10.}

\item{features}{Specific features to label. If specified, show.n is ignored.}

\item{fdr.threshold}{FDR threshold at which reference line is drawn. Default = 0.05.}

\item{label.size}{Label size.}
}
\value{
ggplot handle
}
\description{
Draw volcano plot to visualize differential expression. Uses data.frame output from getDEG(..., return.list = F).
}
\examples{

 df.dat <- getDEG(so.query, return.list = F)
 plt.volcano <-  miko_volcano(df.deg = df.dat)

}
