% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{optimalDS}
\alias{optimalDS}
\title{Cut clustered tree at varying heights and overlay with dendrogram to find optimal parameter set.}
\usage{
optimalDS(tree, d.mat, ...)
}
\arguments{
\item{tree}{h.clust object generated by dist2hclust.}

\item{d.mat}{distance matrix used to generate tree.}

\item{...}{Additional arguments passessed to cutreeHybrid {WGCNA package}}
}
\value{
mColorh, matrix specfying module membership at varying deep split parameter specifications (0:4)
}
\description{
WGCNA::cutreeHybrid is run for varying levels of deepSplit parameter (0:4), and cluster membership is assigned for each parameter set.
}
\examples{

geneTree <- dist2hclust(d.mat)
mColorh <- optimalDS(tree = geneTree, d.mat = d.mat, pamStage = F,cutHeight = 0.99)

plotDendroAndColors(geneTree, mColorh, paste("dpSplt =", 0:4), main = "Group 1", dendroLabels = F)

}
