% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prepExpression}
\alias{prepExpression}
\title{Normalize and scale data within appropriate assay}
\usage{
prepExpression(so)
}
\arguments{
\item{so}{Seurat Object}
}
\value{
Seurat Object
}
\description{
Ensures data are properly normalized and scaled. If intergrated dataset provided, Seurat's NormalizeData, FindVariableFeatures, ScaleData workflow is applied and default assay is set to 'RNA'. Otherwise, if  non-integrated dataset is provided, SCT transform is assumed and default assay is set to 'SCT'. If non-intergrated data have not been process with SCT workflow, see m1.scNormScale() function.
}
