% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prepGeneList}
\alias{prepGeneList}
\title{prep Gene List}
\usage{
prepGeneList(so, global.enviroment)
}
\arguments{
\item{so}{Seurat Object}

\item{global.enviroment}{global.enviroment <- objects()}
}
\value{
Named vector of available genes
}
\description{
Ensure is available and represented correctly.
}
\examples{

gNames.list <- prepGeneList(so, objects())

}
\author{
Nicholas Mikolajewicz
}
