% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_functions.R
\name{prepGeneList}
\alias{prepGeneList}
\title{Set cluster resolution}
\usage{
prepGeneList(so, global.enviroment)
}
\arguments{
\item{so}{Seurat Object}

\item{cluster.resolution}{Seurat Object}
}
\value{
Seurat object
}
\description{
Set 'Seurat_Clusters' metadata entry to quiried cluster resolution
}
