% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prepSeurat2}
\alias{prepSeurat2}
\title{prep Seurat (Extended adaptation of prepSeurat)}
\usage{
prepSeurat2(
  so,
  e2s,
  species,
  resolution = NULL,
  subset.data = NULL,
  subsample = 1,
  M00_subgroup.path = "M00_subgroups.csv",
  terms2drop = NULL
)
}
\arguments{
\item{so}{Seurat objects}

\item{e2s}{ensemble to gene symbol mapping vector. Is a named character vector where names are ENSEMBL and entries are SYMBOLs.}

\item{species}{Species. One of "Mm" or "Hs".}

\item{resolution}{cluster resolution. Numeric [0,Inf] that specifies resolution for data clustering. If requested resolution exists, no new clustering is performed.}

\item{subset.data}{Character or data.frame specifying how to subset seurat object. If character, subsetting information must be provided in M00_subgroup.path file. If data.frame, field column species which meta.data field to subset on, and subgroups column species which subgroup to include within specied field. See scMiko::subsetSeurat() for details.}

\item{subsample}{Numeric [0,1] specifying what fraction of cells to include for analysis. Default is 1. See scMiko::downsampleSeurat() for details.}

\item{M00_subgroup.path}{Path for .csv file containing subsetting information. Read scPipeline documentation for instructions on use.}

\item{terms2drop}{Reduce memory footprint of seurat object by omitting terms that will not be used for current analysis. Supported terms for omission include: "pca", "umap", "ica", "tsne", "nmf", "corr", "gsva", "deg", "counts", "data", "scale", "rna", "sct", "integrated", "graphs".}
}
\value{
list containing prepped Seurat object, default assay, and number of cells in seurat object.
}
\description{
Extended version of original prep Seurat function. In addition to running standard seurat object checks, prepSeurat2 ensures correct species is specified, sets the cluster resolution and subsets and downsamples data. If data are subset or downsampled, then expression values are renormalized and scaled again.
}
\author{
Nicholas Mikolajewicz
}
