% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{QC.violinPlot}
\alias{QC.violinPlot}
\title{QC violin plots}
\usage{
QC.violinPlot(
  so,
  features = c("nFeature_RNA", "nCount_RNA", "percent.mt"),
  group.by = NULL,
  plt.log.flag = T,
  ...
)
}
\arguments{
\item{so}{Seurat Object}

\item{features}{meta data features to plot. Default is c("nFeature_RNA", "nCount_RNA", "percent.mt").}

\item{group.by}{meta data field to group plots by.}

\item{plt.log.flag}{Logical specifying whether data are plotted on log scale. Default is True.}

\item{...}{additional arguments passed to Seurat::VlnPlot(...)}
}
\value{
list of ggplot handles
}
\description{
Number of genes/cell, UMI/cell and mitochondiral content/cell are visualized with violin plots. Two ggplot handles are generated. First contains QC metrics pooled across all cells, while second stratifies dataset by grouping variable.
}
