% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rmDuplicateGenes}
\alias{rmDuplicateGenes}
\title{Remove duplicate genes from Seurat Object}
\usage{
rmDuplicateGenes(so)
}
\arguments{
\item{so}{Seurat Object}
}
\value{
seurat object
}
\description{
Remove duplicate genes from Seurat Object
}
\examples{

so.query <- rmDuplicateGenes(so.query)

}
