% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{runAUC}
\alias{runAUC}
\title{Run AUCell classification}
\usage{
runAUC(
  object,
  genelist,
  assay = DefaultAssay(object),
  n.workers = 1,
  n.repeats = 5,
  n.iterations = 1000,
  posterior.p = 0.9,
  mixture.analysis = T,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{genelist}{Named list of genesets.}

\item{assay}{Assay used for expression matrix.}

\item{n.workers}{Number of workers for parallelized implementation. Default is 1.}

\item{n.repeats}{Number of fitting repeats (best fit taken). Default is 5.}

\item{n.iterations}{Number of fitting iterations per a repeat. Default is 1000.}

\item{posterior.p}{Posterior distribution membership threshold. Default is 0.9 (e.g., p > 0.9 is member).}

\item{mixture.analysis}{logical to perform mixture analysis. Default = T (computationally intensive).}

\item{...}{additional parameters passed to geom_point(...)}
}
\value{
list of results along with ggplot handles visualizing class predictions overlaid on UMAP.
}
\description{
Run AUCell classification to identify active genesets in scRNAseq sample. Wrapper for AUCell package.
}
\examples{

# get genesets
verhaak.df <- geneSets[["Verhaak_CancerCell_2010"]]
verhaak.list <- wideDF2namedList(verhaak.df)

gsc.df <- geneSets[["Richards_NatureCancer_2021_sc"]]
gsc.list <- wideDF2namedList(gsc.df)

neftel.df <- geneSets[["GBM_Hs_Neftel2019"]]
neftel.list <- wideDF2namedList(neftel.df)

verhaak.list <- lapply(verhaak.list, toupper)
gsc.list <- lapply(gsc.list, toupper)
neftel.list <- lapply(neftel.list, toupper)

# classify cells based on provided genesets
v.auc <- runAUC(object = so.query, genelist = verhaak.list, n.workers = 12)
v.auc$plot.auc
v.auc$plot.nm
v.auc$plot.max.score

g.auc <- runAUC(object = so.query, genelist = gsc.list, n.workers = 12)
g.auc$plot.auc
g.auc$plot.nm
g.auc$plot.max.score

n.auc <- runAUC(object = so.query, genelist = neftel.list, n.workers = 12)
n.auc$plot.auc
n.auc$plot.nm
n.auc$plot.max.score

}
\seealso{
\code{\link{AUCell_calcAUC }}
}
\author{
Nicholas Mikolajewicz
}
