% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{runBBKNN}
\alias{runBBKNN}
\title{Run batch balanced KNN}
\usage{
runBBKNN(
  object,
  batch,
  reduction = "pca",
  assay = DefaultAssay(object),
  verbose = T
)
}
\arguments{
\item{object}{Seurat object}

\item{batch}{name of meta data field specifying batch groupings.}

\item{reduction}{reduction used for batch correction. Default is "pca".}

\item{assay}{assay name. Default is DefaultAssay(object).}

\item{verbose}{print progress. Default is T.}
}
\value{
Seurat object with bbKNN-corrected UMAP stored in "b" reduction slot.
}
\description{
This is a wrapper for bbknn (https://github.com/Teichlab/bbknn). BBKNN is a fast and intuitive batch effect removal tool. It generates a BBKNN cell x cell similarity graph that can be used for clustering and UMAP embedding.
}
\examples{
 object <- runBBKNN(object, batch = "sample")

}
\author{
Nicholas Mikolajewicz
}
