% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{runEnrichment}
\alias{runEnrichment}
\title{Wrapper to run  GO enrichment, using fisher enrichment method}
\usage{
runEnrichment(gene.list, gene.universe, species)
}
\arguments{
\item{gene.list}{named list of gene sets to query, where name specify name of gene set, and entry is vector of gene symbols.}

\item{gene.universe}{background genes}

\item{species}{Species: Hs or Mm}
}
\value{
list of data.frames. results.table.p contains unadjusted results, results.table.bh contains BH-adjusted results.
}
\description{
Wrapper to run  GO enrichment, using fisher enrichment method. Builds on TopGo functionality.
}
\examples{

enrichment.list <- runEnrichment(module.list, gene.universe = all.genes, species = "Hs")
results.p <- enrichment.list$results.table.p
results.bh <- enrichment.list$results.table.bh

}
