% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{runHG}
\alias{runHG}
\title{Run hypergeometric gene enrichment analysis.}
\usage{
runHG(
  gene.list,
  gene.universe,
  species,
  pathway.db = "Bader",
  n.workers = 16,
  my.pathway = NULL,
  my.pathway.representation = "ENTREZ",
  min.size = 1,
  max.size = 300,
  e2s = F,
  go.ontology = "BP"
)
}
\arguments{
\item{gene.list}{Named list of genesets to run enrichment on (symbol format).}

\item{gene.universe}{Gene universe (symbol format)}

\item{species}{Species. One of "Mm" (mouse) or "Hs" (human)}

\item{pathway.db}{Name of databse to get annotation genelists from. Default is "Bader". See scMiko::getAnnotationPathways() for options.}

\item{n.workers}{Number of workers for parallelization. Default is 16.}

\item{my.pathway}{Named list of pathways, with each entry containing vector of Entrez IDs. Retrieved internally unless explicitly provided.}

\item{my.pathway.representation}{If my.pathway is provided, species which format, options = "SYMBOL" or "ENTREZ". ENTREZ is Default.}

\item{min.size}{min geneset size. Default is 1.}

\item{max.size}{max geneset size. Default is 300.}

\item{e2s}{entrez to symbol mapping (computationally demanding). Default False.}

\item{go.ontology}{BP, MF or CC. Ignored if pathway.db != "GO".}
}
\description{
Run hypergeometric gene enrichment analysis.
}
\author{
Nicholas Mikolajewicz
}
