% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{runNMF}
\alias{runNMF}
\title{Perform non-negative matrix factorization (NMF)}
\usage{
runNMF(
  object,
  assay = DefaultAssay(object),
  k = 6,
  raster = F,
  n.threads = 2,
  features = NULL,
  feature.min.pct = 0,
  max.iter = 50,
  gene.cutoff = 0.5,
  gene.n = 50,
  sample.name = NULL,
  show.top.n = 10,
  pathway.db = "Bader",
  do.enrichment = T,
  verbose = T,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{assay. Default is DefaultAssay(object).}

\item{k}{Number of NMF gene programs Default is 6.}

\item{raster}{rasterize output plot. Default is F.}

\item{n.threads}{Number of threads to use when running NMF. Default is 2.}

\item{features}{features to run NMF on.}

\item{feature.min.pct}{minimum expressing fraction for feature to run NMF on. Default is 0. Ignored if `features` are specified.}

\item{max.iter}{Maximum iteration of alternating NNLS solutions to H and W}

\item{gene.cutoff}{Feature loading cutoff threshold [0,1]. Default = 0.5. Ignored if gene.n is specified.}

\item{gene.n}{number of genes to return per gene program}

\item{sample.name}{sample name. Default is NULL.}

\item{show.top.n}{number of enrichment terms to show in summary plots. Ignoried if do.enrichment = F.}

\item{pathway.db}{pathway database to use for enrichment analysis. Options are "GO" or "Bader". Default is "Bader". Ignored if do.enrichment = F.}

\item{do.enrichment}{Whether to run enrichment analysis.}

\item{verbose}{Print progress. Default is TRUE.}

\item{...}{additional arguments passed to NNLM::nnmf(...)}
}
\value{
Seurat object with NMF results in reduction slot. Program genes are stored in "misc" slot of NMF reduction slot.
}
\description{
Run non-negative matrix factorization(NMF or NNMF) using sequential coordinate-wise descent or multiplicative updates on Seurat object.
}
\examples{

}
\references{
\url{https://nmikolajewicz.github.io/scMiko/articles/Module_Detection.html}
}
\seealso{
\code{\link{nnmf}}
}
\author{
Nicholas Mikolajewicz
}
