% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{scale_color_miko}
\alias{scale_color_miko}
\title{Gradient color scale}
\usage{
scale_color_miko(low = scales::muted("blue"), high = scales::muted("red"))
}
\arguments{
\item{low}{color for low end of gradient}

\item{high}{color for high end of gradient}
}
\description{
Wrapper for scale_color_gradient2, with preferred default parameters
}
\examples{

 df.dat <- getDEG(so.query, return.list = F)
 plt.volcano <-  miko_volcano(df.deg = df.dat) + scale_color_miko()

}
