% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{scale_fill_miko}
\alias{scale_fill_miko}
\title{Gradient fill scale}
\usage{
scale_fill_miko(low = scales::muted("blue"), high = scales::muted("red"))
}
\arguments{
\item{low}{color for low end of gradient}

\item{high}{color for high end of gradient}
}
\description{
Wrapper for scale_fill_gradient2, with preferred default parameters
}
