% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{scaleFreeNet}
\alias{scaleFreeNet}
\title{Apply scale-free topology transform to shared-nearest neighbor graph}
\usage{
scaleFreeNet(
  object,
  graph_name = "RNA_snn",
  sf = c(seq(1, 5, by = 0.5), seq(6, 10)),
  sf_threshold = 0.9,
  umap_knn = 10,
  n_dim = 30,
  binary_threshold = 0.9,
  n_workers = 1,
  verbose = T
)
}
\arguments{
\item{object}{Seurat object}

\item{sf}{soft thresholds (sf) that are evaluated in search for optimal soft threshold (used to transform SNN to scale-free topology). Default is sf = c(seq(1, 5, by = 0.5), seq(6, 10)).}

\item{sf_threshold}{R2 threshold [0,1] for identifying optimal soft threshold. Default is 0.9 (recommended 0.8-0.9).}

\item{umap_knn}{This determines the number of neighboring points used in local approximations of UMAP manifold structure. Larger values will result in more global structure being preserved at the loss of detailed local structure. In general this parameter should often be in the range 5 to 50. default is 10.}

\item{n_dim}{Number of PC dimensions to use in generating UMAP. Default is PCA.}

\item{binary_threshold}{Numeric threshold used to binarize SNN graph. Default is 0.9.}

\item{n_workers}{Number of workers used for parallel implementation. Default is 1.}

\item{verbose}{Show progress. Default is T.}

\item{graph.name}{name of SNN graph in `object`. Default is "RNA_snn".}
}
\value{
list containing
\itemize{
\item "object" - scale-free SNN and binarized graph are stored in `object` graph slot, and umap embedding is stored in `object` reduction slot.
\item "st.res" - soft threshold optimization results.
}
}
\description{
Apply scale-free topology transform to shared-nearest neighbor graph
}
\examples{

}
\references{
\url{https://nmikolajewicz.github.io/scMiko/articles/Module_Detection.html}
}
\seealso{
\code{\link{runSSN}}
}
\author{
Nicholas Mikolajewicz
}
