% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{scExpression.UMAP}
\alias{scExpression.UMAP}
\title{Cell-level gene expression projected on UMAP}
\usage{
scExpression.UMAP(
  object,
  query.gene,
  x.label = "UMAP 1",
  y.label = "",
  adjust.pt.size = T,
  order.cells = T,
  plot.name = NULL,
  ...
)
}
\arguments{
\item{object}{Seurat Object}

\item{query.gene}{Character. Gene name to plot.}

\item{x.label}{Character. X axis label.}

\item{y.label}{Character. Y axis label.}

\item{adjust.pt.size}{Adjust point size for plotting. Logical.}

\item{order.cells}{Plot cells in order of expression. Logical.}

\item{plot.name}{Character. Plot title.}

\item{...}{additional parameters passed to Seurat::FeaturePlot().}
}
\value{
ggplot handle
}
\description{
UMAP plot with cell-level gene expression for queried gene. Uses Seurat::FeaturePlot().
}
