% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{scMikoReload}
\alias{scMikoReload}
\title{reload scMiko}
\usage{
scMikoReload()
}
\description{
Function that detachs and attaches scMiko package.
}
\examples{

# reload scMiko package
scMikoReload()

}
