% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{scoreGBM}
\alias{scoreGBM}
\title{Annotate glioblastoma (GBM) subtype  based on Neftel 2019 scoring pipeline.}
\usage{
scoreGBM(object, which.species = detectSpecies(object))
}
\arguments{
\item{object}{seurat object}

\item{which.species}{Species, either "Mm" or "Hs".}
}
\description{
Annotate glioblastoma (GBM) subtype  based on Neftel 2019 scoring pipeline.
}
\author{
Nicholas Mikolajewicz
}
