% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{scoreGBM}
\alias{scoreGBM}
\title{Annotate glioblastoma (GBM) subtype  based on Neftel 2019 scoring pipeline.}
\usage{
scoreGBM(object, species = detectSpecies(object), verbose = T)
}
\arguments{
\item{object}{seurat object}

\item{species}{Species, either "Mm" or "Hs".}

\item{verbose}{Print progress. Default is TRUE.}
}
\description{
Annotate glioblastoma (GBM) subtype  based on Neftel 2019 scoring pipeline.
}
\author{
Nicholas Mikolajewicz
}
