% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{setResolution}
\alias{setResolution}
\title{Set cluster resolution}
\usage{
setResolution(so, cluster.resolution)
}
\arguments{
\item{so}{Seurat Object}

\item{cluster.resolution}{Numeric [0, inf] specifying cluster resolution. Values [0.1,1] typically perform well.}
}
\value{
Seurat object
}
\description{
Set 'Seurat_Clusters' metadata entry to specified cluster resolution [0, inf]. See Seurat::FindClusters() for details.
}
\author{
Nicholas Mikolajewicz
}
