% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{springNet}
\alias{springNet}
\title{Generate spring-embedded network from similarity matrix}
\usage{
springNet(
  smat,
  group_by = NULL,
  cex_line = 0.1,
  knn = 5,
  min_edge = quantile(smat, 0.95),
  node_label_size = NULL,
  seed = 1023
)
}
\arguments{
\item{smat}{similarity matrix}

\item{group_by}{group variable used to color network nodes}

\item{cex_line}{edge width}

\item{knn}{k-nearest neighbors. Default is 5. If not specified, defaults to min_edge argument.}

\item{min_edge}{similarity threshold used to display network edge}

\item{node_label_size}{size of node labels. If not specified (i.e., node_label_size = NULL), no label is shown.}

\item{seed}{number used to initialize pseudorandom number generator. Specify to ensure reproducible graphs. Default is 1023.}
}
\value{
ggplot
}
\description{
Generate spring-embedded network from similarity matrix
}
\examples{

smat <- cor(x)
springNet(smat)

}
\author{
Nicholas Mikolajewicz
}
