% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{SSNConnectivity}
\alias{SSNConnectivity}
\title{SSN connectivity plot}
\usage{
SSNConnectivity(
  gene.object,
  gene.list = NULL,
  quantile_threshold = 0.9,
  raster_dpi = 200,
  edge.alpha = 0.015,
  do.label = T,
  label.size = 5,
  verbose = T
)
}
\arguments{
\item{gene.object}{Seurat object (cell x gene) obtained from SSN analysis.}

\item{gene.list}{named list of module features. If omitted, feature-annotated connectivity plot is not generated.}

\item{quantile_threshold}{Quantile threshold for visualized SSN graph edges. Default is 0.9.}

\item{raster_dpi}{prefix added to each named entry in list. Default is "".}

\item{edge.alpha}{alpha [0,1] parameter (i.e., transparency) for network edges. Default is 0.015. Use larger values for less dense networks.}

\item{do.label}{Show module number on network graph. Default is T.}

\item{label.size}{Size of module numbers on network graph. Default is 5.}

\item{verbose}{Print progress. Default is T.}
}
\value{
Returns 2 variants of SSN connectivity plot along with data.frame used to generate plots.
}
\description{
generates SSN connectivity plot used to visualize transcriptomic gene network.
}
\examples{

}
\seealso{
\code{\link{runSSN}} for SSN analysis (gene.object), \code{\link{pruneSSN}} for module features (gene.list)
}
\author{
Nicholas Mikolajewicz
}
