% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{SSNResolution}
\alias{SSNResolution}
\title{Identify optimal cluster resolution of scale-free shared nearest neighbor network (SNN)}
\usage{
SSNResolution(
  object,
  graph,
  target.purity = 0.7,
  start.res = 0.5,
  step.size = 0.05,
  target.level = "global",
  max.iter = 100,
  verbose = F
)
}
\arguments{
\item{object}{Seurat object}

\item{graph}{name of SNN graph in `object`.}

\item{target.purity}{Target purity to find resolution for.}

\item{start.res}{Initial resolution to begin optimization with. Default is 0.5.}

\item{step.size}{Step size between consecutive resolutions to test. Default is 0.05.}

\item{target.level}{Specify whether `target.purity` is achieved at "global" level [i.e., median(purity)] or at "cluster" level. Default is "global".}

\item{max.iter}{Maximum number of iterations. If `max.iter` is reached, non-optimal resolution may be returned.}

\item{verbose}{Show progress. Default is T.}
}
\value{
optimal cluster resolution
}
\description{
Identify optimal cluster resolution of scale-free shared nearest neighbor network (SNN)
}
\examples{

}
\references{
\url{https://nmikolajewicz.github.io/scMiko/articles/Module_Detection.html}
}
\seealso{
\code{\link{neighborPurity}}
}
\author{
Nicholas Mikolajewicz
}
