% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_functions.R
\name{subsetSeurat}
\alias{subsetSeurat}
\title{Subset Seurat Object}
\usage{
subsetSeurat(so, subset.df = NULL)
}
\arguments{
\item{so}{Seurat object}

\item{subset.df}{Data.frame specifying which field to subset on, and which field entires to retain.}
}
\value{
Seurat object
}
\description{
Subset Seurat object according to specific metadata field. Only specified metadata entries are retained, while remaining of data is omitted.
}
\examples{
subset.df <- data.frame(field = "seurat_clusters", subgroups = c(0,2,3,5,6,7,9,15)) # subset tumor population of Renca 4000ds T12 tumor
so <- subsetSeurat(so, subset.df)

}
