% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{updateGeneSets}
\alias{updateGeneSets}
\title{Create/update list of gene sets for scMiko package.}
\usage{
updateGeneSets(input.file, output.file, dir = "", dev.directory.flag = F)
}
\arguments{
\item{input.file}{Excel file (input). Must have ".xlsx" suffix. A character.}

\item{output.file}{Rdata file (output). Must have ".rda" suffix. A character.}

\item{dir}{Directory of input and output file (same folder). A character.}

\item{dev.directory.flag}{Logical indicating whether to use developer specific director. Default is False. If true dir is ignored.}
}
\value{
List of data.frames saved as Rdata file.
}
\description{
Takes genessets stored in Excel sheets, and converts them to dataframes stored in lists.
}
\examples{
\dontrun{
updateGeneSets("geneSets_MASTER_260420update.xlsx", "geneSets.rda", dev.directory.flag = T)
}

}
